+++ Validation/Configuration/python/postValidation_cff.py
@@ -100,7 +100,7 @@ postValidation_gen = cms.Sequence(
 )
 
 postValidationCosmics = cms.Sequence(
-    postProcessorMuonMultiTrack
+    postProcessorMuonTrack
 )
 
 postValidationMiniAOD = cms.Sequence(
