+++ Validation/RecoHI/python/track_selectors_cff.py
@@ -2,13 +2,13 @@ import FWCore.ParameterSet.Config as cms
 
 # pt-selection of reco tracks
 import PhysicsTools.RecoAlgos.recoTrackSelector_cfi
-NEWcutsRecoTrkMuons = PhysicsTools.RecoAlgos.recoTrackSelector_cfi.recoTrackSelector.clone()
-NEWcutsRecoTrkMuons.src = "hiGeneralTracks"
-NEWcutsRecoTrkMuons.quality = []
-NEWcutsRecoTrkMuons.ptMin = 0.0
+cutsRecoTrkMuons = PhysicsTools.RecoAlgos.recoTrackSelector_cfi.recoTrackSelector.clone()
+cutsRecoTrkMuons.src = "hiGeneralTracks"
+cutsRecoTrkMuons.quality = []
+cutsRecoTrkMuons.ptMin = 0.0
 
 # pt-selection of tracking particles
 import PhysicsTools.RecoAlgos.trackingParticleSelector_cfi
-NEWcutsTpMuons = PhysicsTools.RecoAlgos.trackingParticleSelector_cfi.trackingParticleSelector.clone()
-NEWcutsTpMuons.ptMin = 0.0
+cutsTpMuons = PhysicsTools.RecoAlgos.trackingParticleSelector_cfi.trackingParticleSelector.clone()
+cutsTpMuons.ptMin = 0.0
