+++ Validation/RecoMuon/plugins/MuonTrackValidator.cc
@@ -1,15 +1,13 @@
 #include "Validation/RecoMuon/plugins/MuonTrackValidator.h"
 #include "DQMServices/ClientConfig/interface/FitSlicesYTool.h"
-
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
 #include "DataFormats/TrackReco/interface/Track.h"
 #include "DataFormats/TrackReco/interface/TrackFwd.h"
-#include "DataFormats/GsfTrackReco/interface/GsfTrack.h"
-#include "DataFormats/GsfTrackReco/interface/GsfTrackFwd.h"
 #include "SimDataFormats/Track/interface/SimTrackContainer.h"
 #include "SimDataFormats/Vertex/interface/SimVertexContainer.h"
+#include "SimDataFormats/PileupSummaryInfo/interface/PileupSummaryInfo.h"
 #include "SimTracker/TrackerHitAssociation/interface/TrackerHitAssociator.h"
 #include "SimTracker/Records/interface/TrackAssociatorRecord.h"
 #include "TrackingTools/TrajectoryState/interface/FreeTrajectoryState.h"
@@ -26,7 +24,6 @@ using namespace edm;
 
 void MuonTrackValidator::bookHistograms(DQMStore::IBooker& ibooker, edm::Run const&, edm::EventSetup const& setup) {
 
-  int j=0;
   for (unsigned int ww=0;ww<associators.size();ww++){
     for (unsigned int www=0;www<label.size();www++){
 
@@ -52,196 +49,196 @@ void MuonTrackValidator::bookHistograms(DQMStore::IBooker& ibooker, edm::Run con
       std::replace(dirName.begin(), dirName.end(), ':', '_');
       ibooker.setCurrentFolder(dirName);
 
-      setUpVectors();
+      h_tracks.push_back( ibooker.book1D("Ntracks","Number of reconstructed tracks",100,-0.5,99.5) );
+      h_fakes.push_back( ibooker.book1D("Nfakes","Number of fake reco tracks",20,-0.5,19.5) );
+      h_charge.push_back( ibooker.book1D("Ncharge","track charge",3,-1.5,1.5) );
 
-      ibooker.goUp();
-      string subDirName = dirName + "/simulation";
-      ibooker.setCurrentFolder(subDirName);
-      h_ptSIM.push_back( ibooker.book1D("ptSIM", "generated p_{t}", 5500, 0, 110 ) );
-      h_etaSIM.push_back( ibooker.book1D("etaSIM", "generated pseudorapidity", 500, -2.5, 2.5 ) );
-      h_tracksSIM.push_back( ibooker.book1D("tracksSIM","number of simulated tracks",200,-0.5,99.5) );
-      h_vertposSIM.push_back( ibooker.book1D("vertposSIM","Transverse position of sim vertices",100,0.,120.) );
-      
-      ibooker.cd();
-      ibooker.setCurrentFolder(dirName);
-      h_tracks.push_back( ibooker.book1D("tracks","number of reconstructed tracks",200,-0.5,19.5) );
-      h_fakes.push_back( ibooker.book1D("fakes","number of fake reco tracks",20,-0.5,19.5) );
-      h_charge.push_back( ibooker.book1D("charge","charge",3,-1.5,1.5) );
-      h_hits.push_back( ibooker.book1D("hits", "number of hits per track", nintHit,minHit,maxHit ) );
-      h_losthits.push_back( ibooker.book1D("losthits", "number of lost hits per track", nintHit,minHit,maxHit) );
-      h_nchi2.push_back( ibooker.book1D("chi2", "normalized #chi^{2}", 200, 0, 20 ) );
-      h_nchi2_prob.push_back( ibooker.book1D("chi2_prob", "normalized #chi^{2} probability",100,0,1));
+      h_recoeta.push_back( ibooker.book1D("num_reco_eta","N of reco track vs eta",nintEta,minEta,maxEta) );
+      h_assoceta.push_back( ibooker.book1D("num_assoSimToReco_eta","N of associated tracks (simToReco) vs eta",nintEta,minEta,maxEta) );
+      h_assoc2eta.push_back( ibooker.book1D("num_assoRecoToSim_eta","N of associated (recoToSim) tracks vs eta",nintEta,minEta,maxEta) );
+      h_simuleta.push_back( ibooker.book1D("num_simul_eta","N of simulated tracks vs eta",nintEta,minEta,maxEta) );
+      h_misideta.push_back( ibooker.book1D("num_chargemisid_eta","N of associated (simToReco) tracks with charge misID vs eta",nintEta,minEta,maxEta) );
 
-      /// this are needed to calculate efficiency during tha harvesting for the automated validation
-      h_recoeta.push_back( ibooker.book1D("num_reco_eta","N of reco track vs eta",nint,min,max) );
-      h_assoceta.push_back( ibooker.book1D("num_assoc(simToReco)_eta","N of associated tracks (simToReco) vs eta",nint,min,max) );
-      h_assoc2eta.push_back( ibooker.book1D("num_assoc(recoToSim)_eta","N of associated (recoToSim) tracks vs eta",nint,min,max) );
-      h_simuleta.push_back( ibooker.book1D("num_simul_eta","N of simulated tracks vs eta",nint,min,max) );
-      h_recopT.push_back( ibooker.book1D("num_reco_pT","N of reco track vs pT",nintpT,minpT,maxpT) );
-      h_assocpT.push_back( ibooker.book1D("num_assoc(simToReco)_pT","N of associated tracks (simToReco) vs pT",nintpT,minpT,maxpT) );
-      h_assoc2pT.push_back( ibooker.book1D("num_assoc(recoToSim)_pT","N of associated (recoToSim) tracks vs pT",nintpT,minpT,maxpT) );
-      h_simulpT.push_back( ibooker.book1D("num_simul_pT","N of simulated tracks vs pT",nintpT,minpT,maxpT) );
-      //
-      h_recohit.push_back( ibooker.book1D("num_reco_hit","N of reco track vs hit",nintHit,minHit,maxHit) );
-      h_assochit.push_back( ibooker.book1D("num_assoc(simToReco)_hit","N of associated tracks (simToReco) vs hit",nintHit,minHit,maxHit) );
-      h_assoc2hit.push_back( ibooker.book1D("num_assoc(recoToSim)_hit","N of associated (recoToSim) tracks vs hit",nintHit,minHit,maxHit) );
-      h_simulhit.push_back( ibooker.book1D("num_simul_hit","N of simulated tracks vs hit",nintHit,minHit,maxHit) );
-      //
+      h_recopT.push_back( ibooker.book1D("num_reco_pT","N of reco track vs pT",nintPt,minPt,maxPt) );
+      h_assocpT.push_back( ibooker.book1D("num_assoSimToReco_pT","N of associated tracks (simToReco) vs pT",nintPt,minPt,maxPt) );
+      h_assoc2pT.push_back( ibooker.book1D("num_assoRecoToSim_pT","N of associated (recoToSim) tracks vs pT",nintPt,minPt,maxPt) );
+      h_simulpT.push_back( ibooker.book1D("num_simul_pT","N of simulated tracks vs pT",nintPt,minPt,maxPt) );
+      h_misidpT.push_back( ibooker.book1D("num_chargemisid_pT","N of associated (simToReco) tracks with charge misID vs pT",nintPt,minPt,maxPt) );      
+
       h_recophi.push_back( ibooker.book1D("num_reco_phi","N of reco track vs phi",nintPhi,minPhi,maxPhi) );
-      h_assocphi.push_back( ibooker.book1D("num_assoc(simToReco)_phi","N of associated tracks (simToReco) vs phi",nintPhi,minPhi,maxPhi) );
-      h_assoc2phi.push_back( ibooker.book1D("num_assoc(recoToSim)_phi","N of associated (recoToSim) tracks vs phi",nintPhi,minPhi,maxPhi) );
+      h_assocphi.push_back( ibooker.book1D("num_assoSimToReco_phi","N of associated tracks (simToReco) vs phi",nintPhi,minPhi,maxPhi) );
+      h_assoc2phi.push_back( ibooker.book1D("num_assoRecoToSim_phi","N of associated (recoToSim) tracks vs phi",nintPhi,minPhi,maxPhi) );
       h_simulphi.push_back( ibooker.book1D("num_simul_phi","N of simulated tracks vs phi",nintPhi,minPhi,maxPhi) );
+      h_misidphi.push_back( ibooker.book1D("num_chargemisid_phi","N of associated (simToReco) tracks with charge misID vs phi",nintPhi,minPhi,maxPhi) );
+
+      h_recohit.push_back( ibooker.book1D("num_reco_hit","N of reco tracks vs N SimHits",nintNHit,minNHit,maxNHit) );
+      h_assochit.push_back( ibooker.book1D("num_assoSimToReco_hit","N of associated tracks (simToReco) vs N SimHits",nintNHit,minNHit,maxNHit) );
+      h_assoc2hit.push_back( ibooker.book1D("num_assoRecoToSim_hit","N of associated (recoToSim) tracks vs N Rechits",nintNHit,minNHit,maxNHit) );
+      h_simulhit.push_back( ibooker.book1D("num_simul_hit","N of simulated tracks vs N SimHits",nintNHit,minNHit,maxNHit) );
+      h_misidhit.push_back( ibooker.book1D("num_chargemisid_hit","N of associated (recoToSim) tracks with charge misID vs N RecHits",nintNHit,minNHit,maxNHit) );
 
       h_recodxy.push_back( ibooker.book1D("num_reco_dxy","N of reco track vs dxy",nintDxy,minDxy,maxDxy) );
-      h_assocdxy.push_back( ibooker.book1D("num_assoc(simToReco)_dxy","N of associated tracks (simToReco) vs dxy",nintDxy,minDxy,maxDxy) );
-      h_assoc2dxy.push_back( ibooker.book1D("num_assoc(recoToSim)_dxy","N of associated (recoToSim) tracks vs dxy",nintDxy,minDxy,maxDxy) );
+      h_assocdxy.push_back( ibooker.book1D("num_assoSimToReco_dxy","N of associated tracks (simToReco) vs dxy",nintDxy,minDxy,maxDxy) );
+      h_assoc2dxy.push_back( ibooker.book1D("num_assoRecoToSim_dxy","N of associated (recoToSim) tracks vs dxy",nintDxy,minDxy,maxDxy) );
       h_simuldxy.push_back( ibooker.book1D("num_simul_dxy","N of simulated tracks vs dxy",nintDxy,minDxy,maxDxy) );
-      
+      h_misiddxy.push_back( ibooker.book1D("num_chargemisid_dxy","N of associated (simToReco) tracks with charge misID vs dxy",nintDxy,minDxy,maxDxy) );           
       h_recodz.push_back( ibooker.book1D("num_reco_dz","N of reco track vs dz",nintDz,minDz,maxDz) );
-      h_assocdz.push_back( ibooker.book1D("num_assoc(simToReco)_dz","N of associated tracks (simToReco) vs dz",nintDz,minDz,maxDz) );
-      h_assoc2dz.push_back( ibooker.book1D("num_assoc(recoToSim)_dz","N of associated (recoToSim) tracks vs dz",nintDz,minDz,maxDz) );
+      h_assocdz.push_back( ibooker.book1D("num_assoSimToReco_dz","N of associated tracks (simToReco) vs dz",nintDz,minDz,maxDz) );
+      h_assoc2dz.push_back( ibooker.book1D("num_assoRecoToSim_dz","N of associated (recoToSim) tracks vs dz",nintDz,minDz,maxDz) );
       h_simuldz.push_back( ibooker.book1D("num_simul_dz","N of simulated tracks vs dz",nintDz,minDz,maxDz) );
+      h_misiddz.push_back( ibooker.book1D("num_chargemisid_dz","N of associated (simToReco) tracks with charge misID vs dz",nintDz,minDz,maxDz) );
 
-      h_assocvertpos.push_back( ibooker.book1D("num_assoc(simToReco)_vertpos","N of associated tracks (simToReco) vs transverse vert position",nintVertpos,minVertpos,maxVertpos) );
-      h_simulvertpos.push_back( ibooker.book1D("num_simul_vertpos","N of simulated tracks vs transverse vert position",nintVertpos,minVertpos,maxVertpos) );
+      h_assocRpos.push_back( ibooker.book1D("num_assoSimToReco_Rpos","N of associated tracks (simToReco) vs Radius",nintRpos,minRpos,maxRpos) );
+      h_simulRpos.push_back( ibooker.book1D("num_simul_Rpos","N of simulated tracks vs Radius",nintRpos,minRpos,maxRpos) );
 
-      h_assoczpos.push_back( ibooker.book1D("num_assoc(simToReco)_zpos","N of associated tracks (simToReco) vs z vert position",nintZpos,minZpos,maxZpos) );
-      h_simulzpos.push_back( ibooker.book1D("num_simul_zpos","N of simulated tracks vs z vert position",nintZpos,minZpos,maxZpos) );
+      h_assocZpos.push_back( ibooker.book1D("num_assoSimToReco_Zpos","N of associated tracks (simToReco) vs Z",nintZpos,minZpos,maxZpos) );
+      h_simulZpos.push_back( ibooker.book1D("num_simul_Zpos","N of simulated tracks vs Z",nintZpos,minZpos,maxZpos) );
 
-      /////////////////////////////////
+      h_recopu.push_back( ibooker.book1D("num_reco_pu","N of reco track vs pu",nintPU,minPU,maxPU) );
+      h_assocpu.push_back( ibooker.book1D("num_assoSimToReco_pu","N of associated tracks (simToReco) vs pu",nintPU,minPU,maxPU) );
+      h_assoc2pu.push_back( ibooker.book1D("num_assoRecoToSim_pu","N of associated (recoToSim) tracks vs pu",nintPU,minPU,maxPU) );
+      h_simulpu.push_back( ibooker.book1D("num_simul_pu","N of simulated tracks vs pu",nintPU,minPU,maxPU) );
+      h_misidpu.push_back( ibooker.book1D("num_chargemisid_pu","N of associated (simToReco) charge misIDed tracks vs pu",nintPU,minPU,maxPU) );
 
-      h_eta.push_back( ibooker.book1D("eta", "pseudorapidity residue", 1000, -0.1, 0.1 ) );
-      h_pt.push_back( ibooker.book1D("pullPt", "pull of p_{t}", 100, -10, 10 ) );
-      h_pullTheta.push_back( ibooker.book1D("pullTheta","pull of #theta parameter",250,-25,25) );
-      h_pullPhi.push_back( ibooker.book1D("pullPhi","pull of #phi parameter",250,-25,25) );
-      h_pullDxy.push_back( ibooker.book1D("pullDxy","pull of dxy parameter",250,-25,25) );
-      h_pullDz.push_back( ibooker.book1D("pullDz","pull of dz parameter",250,-25,25) );
-      h_pullQoverp.push_back( ibooker.book1D("pullQoverp","pull of qoverp parameter",250,-25,25) );
+      h_nchi2.push_back( ibooker.book1D("chi2","Track normalized #chi^{2}", 80, 0., 20.) );
+      h_nchi2_prob.push_back( ibooker.book1D("chi2prob", "Probability of track normalized #chi^{2}",100,0.,1.) );
+
+      chi2_vs_nhits.push_back( ibooker.book2D("chi2_vs_nhits","#chi^{2} vs nhits",nintNHit,minNHit,maxNHit,20,0.,10.) );
+      chi2_vs_eta.push_back( ibooker.book2D("chi2_vs_eta","chi2_vs_eta",nintEta,minEta,maxEta, 40, 0., 20. ));
+      chi2_vs_phi.push_back( ibooker.book2D("chi2_vs_phi","#chi^{2} vs #phi",nintPhi,minPhi,maxPhi, 40, 0., 20. ) );
+
+      h_nhits.push_back( ibooker.book1D("nhits", "Number of hits per track", nintNHit,minNHit,maxNHit ) );
+      nhits_vs_eta.push_back( ibooker.book2D("nhits_vs_eta","Number of Hits vs eta",nintEta,minEta,maxEta,nintNHit,minNHit,maxNHit) );
+      nhits_vs_phi.push_back( ibooker.book2D("nhits_vs_phi","#hits vs #phi",nintPhi,minPhi,maxPhi,nintNHit,minNHit,maxNHit) );
       
-      if (associators[ww]=="trackAssociatorByChi2"){
-	h_assochi2.push_back( ibooker.book1D("assocChi2","track association #chi^{2}",1000000,0,100000) );
-	h_assochi2_prob.push_back(ibooker.book1D("assocChi2_prob","probability of association #chi^{2}",100,0,1));
-      } else if (associators[ww]=="trackAssociatorByHits"){
-	h_assocFraction.push_back( ibooker.book1D("assocFraction","fraction of shared hits",200,0,2) );
-	h_assocSharedHit.push_back(ibooker.book1D("assocSharedHit","number of shared hits",20,0,20));
+      if (do_MUOhitsPlots) {
+	nDThits_vs_eta.push_back( ibooker.book2D("nDThits_vs_eta","Number of DT hits vs eta",nintEta,minEta,maxEta,nintDTHit,minDTHit,maxDTHit) );
+	nCSChits_vs_eta.push_back( ibooker.book2D("nCSChits_vs_eta","Number of CSC hits vs eta",nintEta,minEta,maxEta,nintCSCHit,minCSCHit,maxCSCHit) );
+	nRPChits_vs_eta.push_back( ibooker.book2D("nRPChits_vs_eta","Number of RPC hits vs eta",nintEta,minEta,maxEta,nintRPCHit,minRPCHit,maxRPCHit) );
+	if (useGEMs_)
+	  nGEMhits_vs_eta.push_back( ibooker.book2D("nGEMhits_vs_eta","Number of GEM hits vs eta",nintEta,minEta,maxEta,nintNHit,minNHit,maxNHit) );
+	if(useME0_)
+	  nME0hits_vs_eta.push_back( ibooker.book2D("nME0hits_vs_eta","Number of ME0 hits vs eta",nintEta,minEta,maxEta,nintNHit,minNHit,maxNHit) );
+
       }
-      
-      chi2_vs_nhits.push_back( ibooker.book2D("chi2_vs_nhits","#chi^{2} vs nhits",25,0,25,100,0,10) );
-      h_chi2meanhitsh.push_back( ibooker.bookProfile("chi2mean_vs_nhits","mean #chi^{2} vs nhits",25,0,25,100,0,10) );
 
-      etares_vs_eta.push_back( ibooker.book2D("etares_vs_eta","etaresidue vs eta",nint,min,max,200,-0.1,0.1) );
-      nrec_vs_nsim.push_back( ibooker.book2D("nrec_vs_nsim","nrec vs nsim",20,-0.5,19.5,20,-0.5,19.5) );
+      if (do_TRKhitsPlots) {
+	nTRK_LayersWithMeas_vs_eta.push_back(ibooker.book2D("nTRK_LayersWithMeas_vs_eta","# TRK Layers with measurement vs eta",
+							    nintEta,minEta,maxEta,nintLayers,minLayers,maxLayers) );
+	nPixel_LayersWithMeas_vs_eta.push_back(ibooker.book2D("nPixel_LayersWithMeas_vs_eta","Number of Pixel Layers with measurement vs eta",
+							      nintEta,minEta,maxEta,nintPixels,minPixels,maxPixels) );
+	h_nmisslayers_inner.push_back( ibooker.book1D("nTRK_misslayers_inner", "Number of missing inner TRK layers", nintLayers,minLayers,maxLayers ) );
+	h_nmisslayers_outer.push_back( ibooker.book1D("nTRK_misslayers_outer", "Number of missing outer TRK layers", nintLayers,minLayers,maxLayers ) );
+	h_nlosthits.push_back( ibooker.book1D("nlosthits", "Number of lost hits per track", 6,-0.5,5.5 ) );
+	nlosthits_vs_eta.push_back( ibooker.book2D("nlosthits_vs_eta","Number of lost hits per track vs eta",nintEta,minEta,maxEta,6,-0.5,5.5) );
+      }
 
-      chi2_vs_eta.push_back( ibooker.book2D("chi2_vs_eta","chi2_vs_eta",nint,min,max, 200, 0, 20 ));
-      h_chi2meanh.push_back( ibooker.bookProfile("chi2mean","mean #chi^{2} vs #eta",nint,min,max, 200, 0, 20) );
-      chi2_vs_phi.push_back( ibooker.book2D("chi2_vs_phi","#chi^{2} vs #phi",nintPhi,minPhi,maxPhi, 200, 0, 20 ) );
-      h_chi2mean_vs_phi.push_back( ibooker.bookProfile("chi2mean_vs_phi","mean of #chi^{2} vs #phi",nintPhi,minPhi,maxPhi, 200, 0, 20) );
+      ptres_vs_eta.push_back(ibooker.book2D("ptres_vs_eta","p_{T} Relative Residual vs #eta",
+					    nintEta,minEta,maxEta, ptRes_nbin, ptRes_rangeMin, ptRes_rangeMax));
+      ptres_vs_phi.push_back( ibooker.book2D("ptres_vs_phi","p_{T} Relative Residual vs #phi",
+					     nintPhi,minPhi,maxPhi, ptRes_nbin, ptRes_rangeMin, ptRes_rangeMax));
+      ptres_vs_pt.push_back(ibooker.book2D("ptres_vs_pt","p_{T} Relative Residual vs p_{T}",
+					   nintPt,minPt,maxPt, ptRes_nbin, ptRes_rangeMin, ptRes_rangeMax));
+      h_ptpull.push_back( ibooker.book1D("ptpull", "p_{T} Pull", 100, -10., 10.) );
+      ptpull_vs_eta.push_back(ibooker.book2D("ptpull_vs_eta","p_{T} Pull vs #eta",nintEta,minEta,maxEta,100,-10.,10.));
+      ptpull_vs_phi.push_back(ibooker.book2D("ptpull_vs_phi","p_{T} Pull vs #phi",nintPhi,minPhi,maxPhi,100,-10.,10.));
+      h_qoverppull.push_back( ibooker.book1D("qoverppull","q/p Pull", 100, -10., 10.) );
 
-      nhits_vs_eta.push_back( ibooker.book2D("nhits_vs_eta","nhits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      nDThits_vs_eta.push_back( ibooker.book2D("nDThits_vs_eta","# DT hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      nCSChits_vs_eta.push_back( ibooker.book2D("nCSChits_vs_eta","# CSC hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      nRPChits_vs_eta.push_back( ibooker.book2D("nRPChits_vs_eta","# RPC hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      if(useGEMs_) nGEMhits_vs_eta.push_back( ibooker.book2D("nGEMhits_vs_eta","# GEM hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      if(useME0_) nME0hits_vs_eta.push_back( ibooker.book2D("nME0hits_vs_eta","# ME0 hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
+      h_etaRes.push_back( ibooker.book1D("etaRes", "#eta residual", etaRes_nbin, etaRes_rangeMin, etaRes_rangeMax) );
+      etares_vs_eta.push_back( ibooker.book2D("etares_vs_eta","#eta Residual vs #eta",
+					      nintEta,minEta,maxEta, etaRes_nbin, etaRes_rangeMin, etaRes_rangeMax) );
 
-      h_DThits_eta.push_back( ibooker.bookProfile("DThits_eta","mean # DT hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      h_CSChits_eta.push_back( ibooker.bookProfile("CSChits_eta","mean # CSC hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      h_RPChits_eta.push_back( ibooker.bookProfile("RPChits_eta","mean # RPC hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      if(useGEMs_) h_GEMhits_eta.push_back( ibooker.bookProfile("GEMhits_eta","mean # GEM hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      if(useME0_) h_ME0hits_eta.push_back( ibooker.bookProfile("ME0hits_eta","mean # ME0 hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
+      thetaCotres_vs_eta.push_back(ibooker.book2D("thetaCotres_vs_eta","cot(#theta) Residual vs #eta",
+						  nintEta,minEta,maxEta,cotThetaRes_nbin, cotThetaRes_rangeMin, cotThetaRes_rangeMax));
+      thetaCotres_vs_pt.push_back(ibooker.book2D("thetaCotres_vs_pt","cot(#theta) Residual vs p_{T}",
+						 nintPt,minPt,maxPt, cotThetaRes_nbin, cotThetaRes_rangeMin, cotThetaRes_rangeMax));
+      h_thetapull.push_back( ibooker.book1D("thetapull","#theta Pull",100,-10.,10.) );
+      thetapull_vs_eta.push_back(ibooker.book2D("thetapull_vs_eta","#theta Pull vs #eta",nintEta,minEta,maxEta,100,-10,10));
+      thetapull_vs_phi.push_back(ibooker.book2D("thetapull_vs_phi","#theta Pull vs #phi",nintPhi,minPhi,maxPhi,100,-10,10));
 
-      h_hits_eta.push_back( ibooker.bookProfile("hits_eta","mean #hits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      nhits_vs_phi.push_back( ibooker.book2D("nhits_vs_phi","#hits vs #phi",nintPhi,minPhi,maxPhi,nintHit,minHit,maxHit) );
-      h_hits_phi.push_back( ibooker.bookProfile("hits_phi","mean #hits vs #phi",nintPhi,minPhi,maxPhi, nintHit,minHit,maxHit) );
+      phires_vs_eta.push_back(ibooker.book2D("phires_vs_eta","#phi Residual vs #eta",
+					     nintEta,minEta,maxEta, phiRes_nbin, phiRes_rangeMin, phiRes_rangeMax));
+      phires_vs_pt.push_back(ibooker.book2D("phires_vs_pt","#phi Residual vs p_{T}",
+						nintPt,minPt,maxPt, phiRes_nbin, phiRes_rangeMin, phiRes_rangeMax));
+      phires_vs_phi.push_back(ibooker.book2D("phires_vs_phi","#phi Residual vs #phi",
+					     nintPhi,minPhi,maxPhi,phiRes_nbin, phiRes_rangeMin, phiRes_rangeMax));
+      h_phipull.push_back( ibooker.book1D("phipull","#phi Pull",100,-10.,10.) );
+      phipull_vs_eta.push_back(ibooker.book2D("phipull_vs_eta","#phi Pull vs #eta",nintEta,minEta,maxEta,100,-10,10));
+      phipull_vs_phi.push_back(ibooker.book2D("phipull_vs_phi","#phi Pull vs #phi",nintPhi,minPhi,maxPhi,100,-10,10));
 
-      nlosthits_vs_eta.push_back( ibooker.book2D("nlosthits_vs_eta","nlosthits vs eta",nint,min,max,nintHit,minHit,maxHit) );
-      h_losthits_eta.push_back( ibooker.bookProfile("losthits_eta","losthits_eta",nint,min,max,nintHit,minHit,maxHit) );
+      dxyres_vs_eta.push_back(ibooker.book2D("dxyres_vs_eta","dxy Residual vs #eta",
+					     nintEta,minEta,maxEta,dxyRes_nbin, dxyRes_rangeMin, dxyRes_rangeMax));
+      dxyres_vs_pt.push_back( ibooker.book2D("dxyres_vs_pt","dxy Residual vs p_{T}",
+						 nintPt,minPt,maxPt,dxyRes_nbin, dxyRes_rangeMin, dxyRes_rangeMax));
+      h_dxypull.push_back( ibooker.book1D("dxypull","dxy Pull",100,-10.,10.) );
+      dxypull_vs_eta.push_back(ibooker.book2D("dxypull_vs_eta","dxy Pull vs #eta",nintEta,minEta,maxEta,100,-10,10));
 
-      ptres_vs_eta.push_back(ibooker.book2D("ptres_vs_eta","ptres_vs_eta",nint,min,max, ptRes_nbin, ptRes_rangeMin, ptRes_rangeMax));
-      ptres_vs_phi.push_back( ibooker.book2D("ptres_vs_phi","p_{t} res vs #phi",nintPhi,minPhi,maxPhi, ptRes_nbin, ptRes_rangeMin, ptRes_rangeMax));
-      ptres_vs_pt.push_back(ibooker.book2D("ptres_vs_pt","ptres_vs_pt",nintpT,minpT,maxpT, ptRes_nbin, ptRes_rangeMin, ptRes_rangeMax));
+      dzres_vs_eta.push_back(ibooker.book2D("dzres_vs_eta","dz Residual vs #eta",
+					    nintEta,minEta,maxEta,dzRes_nbin, dzRes_rangeMin, dzRes_rangeMax));
+      dzres_vs_pt.push_back(ibooker.book2D("dzres_vs_pt","dz Residual vs p_{T}",
+					       nintPt,minPt,maxPt,dzRes_nbin, dzRes_rangeMin, dzRes_rangeMax));
+      h_dzpull.push_back( ibooker.book1D("dzpull","dz Pull",100,-10.,10.) );
+      dzpull_vs_eta.push_back(ibooker.book2D("dzpull_vs_eta","dz Pull vs #eta",nintEta,minEta,maxEta,100,-10,10));
 
-      cotThetares_vs_eta.push_back(ibooker.book2D("cotThetares_vs_eta","cotThetares_vs_eta",nint,min,max,cotThetaRes_nbin, cotThetaRes_rangeMin, cotThetaRes_rangeMax));
-      cotThetares_vs_pt.push_back(ibooker.book2D("cotThetares_vs_pt","cotThetares_vs_pt",nintpT,minpT,maxpT, cotThetaRes_nbin, cotThetaRes_rangeMin, cotThetaRes_rangeMax));
-
-      phires_vs_eta.push_back(ibooker.book2D("phires_vs_eta","phires_vs_eta",nint,min,max, phiRes_nbin, phiRes_rangeMin, phiRes_rangeMax));
-      phires_vs_pt.push_back(ibooker.book2D("phires_vs_pt","phires_vs_pt",nintpT,minpT,maxpT, phiRes_nbin, phiRes_rangeMin, phiRes_rangeMax));
-      phires_vs_phi.push_back(ibooker.book2D("phires_vs_phi","#phi res vs #phi",nintPhi,minPhi,maxPhi,phiRes_nbin, phiRes_rangeMin, phiRes_rangeMax));
-
-      dxyres_vs_eta.push_back(ibooker.book2D("dxyres_vs_eta","dxyres_vs_eta",nint,min,max,dxyRes_nbin, dxyRes_rangeMin, dxyRes_rangeMax));
-      dxyres_vs_pt.push_back( ibooker.book2D("dxyres_vs_pt","dxyres_vs_pt",nintpT,minpT,maxpT,dxyRes_nbin, dxyRes_rangeMin, dxyRes_rangeMax));
-
-      dzres_vs_eta.push_back(ibooker.book2D("dzres_vs_eta","dzres_vs_eta",nint,min,max,dzRes_nbin, dzRes_rangeMin, dzRes_rangeMax));
-      dzres_vs_pt.push_back(ibooker.book2D("dzres_vs_pt","dzres_vs_pt",nintpT,minpT,maxpT,dzRes_nbin, dzRes_rangeMin, dzRes_rangeMax));
-
-      ptmean_vs_eta_phi.push_back(ibooker.bookProfile2D("ptmean_vs_eta_phi","mean p_{t} vs #eta and #phi",nintPhi,minPhi,maxPhi,nint,min,max,1000,0,1000));
-      phimean_vs_eta_phi.push_back(ibooker.bookProfile2D("phimean_vs_eta_phi","mean #phi vs #eta and #phi",nintPhi,minPhi,maxPhi,nint,min,max,nintPhi,minPhi,maxPhi));
-
-      //pulls of track params vs eta: to be used with fitslicesytool
-      dxypull_vs_eta.push_back(ibooker.book2D("dxypull_vs_eta","dxypull_vs_eta",nint,min,max,100,-10,10));
-      ptpull_vs_eta.push_back(ibooker.book2D("ptpull_vs_eta","ptpull_vs_eta",nint,min,max,100,-10,10));
-      dzpull_vs_eta.push_back(ibooker.book2D("dzpull_vs_eta","dzpull_vs_eta",nint,min,max,100,-10,10));
-      phipull_vs_eta.push_back(ibooker.book2D("phipull_vs_eta","phipull_vs_eta",nint,min,max,100,-10,10));
-      thetapull_vs_eta.push_back(ibooker.book2D("thetapull_vs_eta","thetapull_vs_eta",nint,min,max,100,-10,10));
-
-      //pulls of track params vs phi
-      ptpull_vs_phi.push_back(ibooker.book2D("ptpull_vs_phi","p_{t} pull vs #phi",nintPhi,minPhi,maxPhi,100,-10,10));
-      phipull_vs_phi.push_back(ibooker.book2D("phipull_vs_phi","#phi pull vs #phi",nintPhi,minPhi,maxPhi,100,-10,10));
-      thetapull_vs_phi.push_back(ibooker.book2D("thetapull_vs_phi","#theta pull vs #phi",nintPhi,minPhi,maxPhi,100,-10,10));
-
-      nrecHit_vs_nsimHit_sim2rec.push_back( ibooker.book2D("nrecHit_vs_nsimHit_sim2rec","nrecHit vs nsimHit (Sim2RecAssoc)",nintHit,minHit,maxHit, nintHit,minHit,maxHit ));
-      nrecHit_vs_nsimHit_rec2sim.push_back( ibooker.book2D("nrecHit_vs_nsimHit_rec2sim","nrecHit vs nsimHit (Rec2simAssoc)",nintHit,minHit,maxHit, nintHit,minHit,maxHit ));
-      
+      nRecHits_vs_nSimHits.push_back( ibooker.book2D("nRecHits_vs_nSimHits","nRecHits vs nSimHits",nintNHit,minNHit,maxNHit, nintNHit,minNHit,maxNHit )); 
+     
       if (MABH) {
 	h_PurityVsQuality.push_back
 	  (ibooker.book2D("PurityVsQuality","Purity vs Quality (MABH)",20,0.01,1.01,20,0.01,1.01) );
-	h_assoceta_Quality05.push_back
-	  (ibooker.book1D("num_assoc(simToReco)_eta_Q05","N of associated tracks (simToReco) vs eta (Quality>0.5)",nint,min,max) );
-	h_assoceta_Quality075.push_back
-	  (ibooker.book1D("num_assoc(simToReco)_eta_Q075","N of associated tracks (simToReco) vs eta (Quality>0.75)",nint,min,max) );
-	h_assocpT_Quality05.push_back
-	  (ibooker.book1D("num_assoc(simToReco)_pT_Q05","N of associated tracks (simToReco) vs pT (Quality>0.5)",nintpT,minpT,maxpT) );
-	h_assocpT_Quality075.push_back
-	  (ibooker.book1D("num_assoc(simToReco)_pT_Q075","N of associated tracks (simToReco) vs pT (Quality>0.75)",nintpT,minpT,maxpT) );
-	h_assocphi_Quality05.push_back
-	  (ibooker.book1D("num_assoc(simToReco)_phi_Q05","N of associated tracks (simToReco) vs phi (Quality>0.5)",nintPhi,minPhi,maxPhi) );
-	h_assocphi_Quality075.push_back
-	  (ibooker.book1D("num_assoc(simToReco)_phi_Q075","N of associated tracks (simToReco) vs phi (Quality>0.75)",nintPhi,minPhi,maxPhi) );
       }
 
-      if(useLogPt){
-	BinLogX(dzres_vs_pt[j]->getTH2F());
-	BinLogX(dxyres_vs_pt[j]->getTH2F());
-	BinLogX(phires_vs_pt[j]->getTH2F());
-	BinLogX(cotThetares_vs_pt[j]->getTH2F());
-	BinLogX(ptres_vs_pt[j]->getTH2F());
-	BinLogX(h_recopT[j]->getTH1F());
-	BinLogX(h_assocpT[j]->getTH1F());
-	BinLogX(h_assoc2pT[j]->getTH1F());
-	BinLogX(h_simulpT[j]->getTH1F());
-	if (MABH) {
-	  BinLogX(h_assocpT_Quality05[j]->getTH1F());
-	  BinLogX(h_assocpT_Quality075[j]->getTH1F());
-	}
-	j++;
+      if (associators[ww]=="trackAssociatorByChi2"){
+	h_assochi2.push_back( ibooker.book1D("assocChi2","track association #chi^{2}",1000,0.,100.) );
+	h_assochi2_prob.push_back(ibooker.book1D("assocChi2_prob","probability of association #chi^{2}",100,0.,1.));
+      } else if (associators[ww]=="trackAssociatorByHits"){
+	h_assocFraction.push_back( ibooker.book1D("assocFraction","fraction of shared hits",22,0.,1.1) );
+	h_assocSharedHit.push_back(ibooker.book1D("assocSharedHit","number of shared hits",41,-0.5,40.5));
+      }
+      
+      if (useLogPt) {
+	BinLogX(h_simulpT.back()->getTH1F());
+	BinLogX(h_assocpT.back()->getTH1F());
+	BinLogX(h_recopT.back()->getTH1F());
+	BinLogX(h_assoc2pT.back()->getTH1F());
+	BinLogX(h_misidpT.back()->getTH1F());
+
+	BinLogX(phires_vs_pt.back()->getTH2F());
+	BinLogX(thetaCotres_vs_pt.back()->getTH2F());
+	BinLogX(dxyres_vs_pt.back()->getTH2F());
+	BinLogX(dzres_vs_pt.back()->getTH2F());
+	BinLogX(ptres_vs_pt.back()->getTH2F());
       }
 
-    }
-  }
+    }   //for (unsigned int www=0;www<label.size();www++)
+  }     //for (unsigned int ww=0;ww<associators.size();ww++)
+
 }
 
-void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup& setup){
+void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup& setup) {
   using namespace reco;
 
   edm::LogInfo("MuonTrackValidator") << "\n====================================================" << "\n"
 				     << "Analyzing new event" << "\n"
 				     << "====================================================\n" << "\n";
 
+  edm::Handle< std::vector<PileupSummaryInfo> > puinfoH;
+  int PU_NumInteractions(-1);
+
   edm::ESHandle<ParametersDefinerForTP> Lhc_parametersDefinerTP;
   std::unique_ptr<ParametersDefinerForTP> Cosmic_parametersDefinerTP;
 
   if(parametersDefiner=="LhcParametersDefinerForTP") {
     setup.get<TrackAssociatorRecord>().get(parametersDefiner, Lhc_parametersDefinerTP);
+
+    // PileupSummaryInfo is contained only in collision events
+    event.getByToken(pileupinfo_Token,puinfoH);
+    for (std::vector<PileupSummaryInfo>::const_iterator puInfoIt = puinfoH->begin(); puInfoIt != puinfoH->end(); ++puInfoIt) {
+      if (puInfoIt->getBunchCrossing()==0) {
+	PU_NumInteractions = puInfoIt->getPU_NumInteractions();
+	break;
+      }
+    }
+
   }
   else if(parametersDefiner=="CosmicParametersDefinerForTP") {
     edm::ESHandle<CosmicParametersDefinerForTP>  _Cosmic_parametersDefinerTP;
@@ -275,82 +272,76 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
   std::vector<const reco::TrackToTrackingParticleAssociator*> associator;
   if (UseAssociators) {
     edm::Handle<reco::TrackToTrackingParticleAssociator> theAssociator;
-    for (unsigned int w=0;w<associators.size();w++) {
-      event.getByLabel(associators[w],theAssociator);
+    for (auto const& associatorName : associators) {
+      event.getByLabel(associatorName,theAssociator);
       associator.push_back( theAssociator.product() );
     }
   }
   
   int w=0;
-  for (unsigned int ww=0;ww<associators.size();ww++){
-    for (unsigned int www=0;www<label.size();www++){
+  for (unsigned int ww=0;ww<associators.size();ww++) {
+    for (unsigned int www=0;www<label.size();www++) {
       //
       //get collections from the event
       //
       edm::Handle<edm::View<Track> > trackCollection;
+      unsigned int trackCollectionSize = 0;
 
       reco::RecoToSimCollection recSimColl;
       reco::SimToRecoCollection simRecColl;
-      unsigned int trackCollectionSize = 0;
 
-      if(!event.getByToken(track_Collection_Token[www], trackCollection)&&ignoremissingtkcollection_) {
-
+      // account for missing track collections (HLT case)
+      if (!event.getByToken(track_Collection_Token[www], trackCollection) && ignoremissingtkcollection_) {
 	recSimColl.post_insert();
 	simRecColl.post_insert();
-	
       }
 
+      //associate tracks to TrackingParticles
       else {
-
 	trackCollectionSize = trackCollection->size();
 
-	//associate tracks
-	if(UseAssociators){
+	if(UseAssociators) {
 	  edm::LogVerbatim("MuonTrackValidator") << "Analyzing "
-						 << label[www].process()<<":"
-						 << label[www].label()<<":"
-						 << label[www].instance()<<" with "
-						 << associators[ww].c_str() <<"\n";
+						    << label[www].process()<<":"
+						    << label[www].label()<<":"
+						    << label[www].instance()<<" with "
+						    << associators[ww].c_str() <<"\n";
 	  
 	  LogTrace("MuonTrackValidator") << "Calling associateRecoToSim method" << "\n";
-	  recSimColl=associator[ww]->associateRecoToSim(trackCollection,
-							TPCollectionHfake);
+	  recSimColl = std::move(associator[ww]->associateRecoToSim(trackCollection,
+								    TPCollectionHfake));
 	  LogTrace("MuonTrackValidator") << "Calling associateSimToReco method" << "\n";
-	  simRecColl=associator[ww]->associateSimToReco(trackCollection,
-							TPCollectionHeff);
-	}
-	else{
+	  simRecColl = std::move(associator[ww]->associateSimToReco(trackCollection,
+								    TPCollectionHeff));
+	} else {
 	  edm::LogVerbatim("MuonTrackValidator") << "Analyzing "
-						 << label[www].process()<<":"
-						 << label[www].label()<<":"
-						 << label[www].instance()<<" with "
-						 << associatormap.process()<<":"
-						 << associatormap.label()<<":"
-						 << associatormap.instance()<<"\n";
+						    << label[www].process()<<":"
+						    << label[www].label()<<":"
+						    << label[www].instance()<<" with "
+						    << associatormap.process()<<":"
+						    << associatormap.label()<<":"
+						    << associatormap.instance()<<"\n";
 	  
 	  Handle<reco::SimToRecoCollection > simtorecoCollectionH;
 	  event.getByToken(simToRecoCollection_Token,simtorecoCollectionH);
-	  simRecColl= *(simtorecoCollectionH.product());
+	  simRecColl = *simtorecoCollectionH.product();
 	  
 	  Handle<reco::RecoToSimCollection > recotosimCollectionH;
 	  event.getByToken(recoToSimCollection_Token,recotosimCollectionH);
-	  recSimColl= *(recotosimCollectionH.product());
-	}
-	
+	  recSimColl = *recotosimCollectionH.product();
+	}	
       }
-      
+       
       //
       //fill simulation histograms
-      //compute number of tracks per eta interval
       //
       edm::LogVerbatim("MuonTrackValidator") << "\n# of TrackingParticles: " << tPCeff.size() << "\n";
       int ats = 0;
       int st = 0;
       for (TrackingParticleCollection::size_type i=0; i<tPCeff.size(); i++){
 	bool TP_is_matched = false;
+        bool isChargeOK = true;
 	double quality = 0.;
-	bool Quality05 = false;
-	bool Quality075 = false;
 	
 	TrackingParticleRef tpr(TPCollectionHeff, i);
 	TrackingParticle* tp = const_cast<TrackingParticle*>(tpr.get());
@@ -388,142 +379,15 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 					       <<", \t vertexTP: radius = "<<sqrt(vertexTP.perp2())<<  ",  z = "<<vertexTP.z() <<"\n";
 	st++;
 
-	h_ptSIM[w]->Fill(sqrt(momentumTP.perp2()));
-	h_etaSIM[w]->Fill(momentumTP.eta());
-	h_vertposSIM[w]->Fill(sqrt(vertexTP.perp2()));
-	
-	std::vector<std::pair<RefToBase<Track>, double> > rt;
-	if(simRecColl.find(tpr) != simRecColl.end()){
-	  rt = (std::vector<std::pair<RefToBase<Track>, double> >) simRecColl[tpr];
-	  if (!rt.empty()) {
-	    RefToBase<Track> assoc_recoTrack = rt.begin()->first;
-	    edm::LogVerbatim("MuonTrackValidator")<<"-----------------------------associated Track #"<<assoc_recoTrack.key();
-	    TP_is_matched = true;
-	    ats++;
-	    quality = rt.begin()->second;
-	    edm::LogVerbatim("MuonTrackValidator") << "TrackingParticle #" <<tpr.key()
-						   << " with pt=" << sqrt(momentumTP.perp2())
-						   << " associated with quality:" << quality <<"\n";
-	    if (MABH) {
-	      if (quality > 0.75) {
-		Quality075 = true;
-		Quality05 = true;
-	      }
-	      else if (quality > 0.5) {
-		Quality05 = true;
-	      }
-	    }	
-	  }
-	}else{
-	  edm::LogVerbatim("MuonTrackValidator")
-	    << "TrackingParticle #" << tpr.key()
-	    << " with pt,eta,phi: "
-	    << sqrt(momentumTP.perp2()) << " , "
-	    << momentumTP.eta() << " , "
-	    << momentumTP.phi() << " , "
-	    << " NOT associated to any reco::Track" << "\n";
-	}
-	
-	for (unsigned int f=0; f<etaintervals[w].size()-1; f++){
-	  if (getEta(momentumTP.eta())>etaintervals[w][f]&&
-	      getEta(momentumTP.eta())<etaintervals[w][f+1]) {
-	    totSIMeta[w][f]++;
-	    if (TP_is_matched) {
-	      totASSeta[w][f]++;
-	      
-	      if (MABH) {
-		if (Quality075) {
-		  totASSeta_Quality075[w][f]++;
-		  totASSeta_Quality05[w][f]++;
-		}
-		else if (Quality05) {
-		  totASSeta_Quality05[w][f]++;
-		}
-	      }
-	    }
-	  }
-	} // END for (unsigned int f=0; f<etaintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<phiintervals[w].size()-1; f++){
-	  if (momentumTP.phi() > phiintervals[w][f]&&
-	      momentumTP.phi() <phiintervals[w][f+1]) {
-	    totSIM_phi[w][f]++;
-	    if (TP_is_matched) {
-	      totASS_phi[w][f]++;
-	      
-	      if (MABH) {
-		if (Quality075) {
-		  totASS_phi_Quality075[w][f]++;
-		  totASS_phi_Quality05[w][f]++;
-		}
-		else if (Quality05) {
-		  totASS_phi_Quality05[w][f]++;
-		}
-	      }
-	    }
-	  }
-	} // END for (unsigned int f=0; f<phiintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<pTintervals[w].size()-1; f++){
-          if (getPt(sqrt(momentumTP.perp2()))>pTintervals[w][f]&&
-              getPt(sqrt(momentumTP.perp2()))<pTintervals[w][f+1]) {
-            totSIMpT[w][f]++;
-	    if (TP_is_matched) {
-	      totASSpT[w][f]++;
-	      
-	      if (MABH) {
-		if (Quality075) {
-		  totASSpT_Quality075[w][f]++;
-		  totASSpT_Quality05[w][f]++;
-		}
-		else if (Quality05) {
-		  totASSpT_Quality05[w][f]++;
-		}
-	      }
-	    }
-	  }
-	} // END for (unsigned int f=0; f<pTintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<dxyintervals[w].size()-1; f++){
-	  if (dxySim>dxyintervals[w][f]&&
-	      dxySim<dxyintervals[w][f+1]) {
-	    totSIM_dxy[w][f]++;
-	    if (TP_is_matched) {
-	      totASS_dxy[w][f]++;
-	    }
-	  }
-	} // END for (unsigned int f=0; f<dxyintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<dzintervals[w].size()-1; f++){
-	  if (dzSim>dzintervals[w][f]&&
-	      dzSim<dzintervals[w][f+1]) {
-	    totSIM_dz[w][f]++;
-	    if (TP_is_matched) {
-	      totASS_dz[w][f]++;
-	    }
-	  }
-	} // END for (unsigned int f=0; f<dzintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<vertposintervals[w].size()-1; f++){
-	  if (sqrt(vertexTP.perp2())>vertposintervals[w][f]&&
-	      sqrt(vertexTP.perp2())<vertposintervals[w][f+1]) {
-	    totSIM_vertpos[w][f]++;
-	    if (TP_is_matched) {
-	      totASS_vertpos[w][f]++;
-	    }
-	  }
-	} // END for (unsigned int f=0; f<vertposintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<zposintervals[w].size()-1; f++){
-	  if (vertexTP.z()>zposintervals[w][f]&&
-	      vertexTP.z()<zposintervals[w][f+1]) {
-	    totSIM_zpos[w][f]++;
-	    if (TP_is_matched) {
-	      totASS_zpos[w][f]++;
-	    }
-	  }
-	} // END for (unsigned int f=0; f<zposintervals[w].size()-1; f++){
-	
+	double TPeta = momentumTP.eta();
+	double xTPeta = getEta(TPeta);            // may be |eta| in histos according to useFabsEta
+        double TPpt = sqrt(momentumTP.perp2());
+	double xTPpt = getPt(TPpt);               // may be 1/pt in histos according to useInvPt
+        double TPphi = momentumTP.phi();
+	double TPrpos = sqrt(vertexTP.perp2());
+	double TPzpos = vertexTP.z();
+
+	// Number of counted SimHits depend on the selection of tracker and muon detectors (via cfg parameters) 
 	int nSimHits = 0;
 	if (usetracker && usemuon) {
 	  nSimHits= tpr.get()->numberOfHits();
@@ -534,20 +398,93 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 	else if (usetracker && !usemuon) {
 	  nSimHits=tpr.get()->numberOfTrackerHits();
 	}
-	
-        int tmp = std::min(nSimHits,int(maxHit-1));
 	edm::LogVerbatim("MuonTrackValidator") << "\t N simhits = "<< nSimHits<<"\n";
-	
-	totSIM_hit[w][tmp]++;
-	if (TP_is_matched) totASS_hit[w][tmp]++;
-	
-	if (TP_is_matched)	
-	  {
-	    RefToBase<Track> assoctrack = rt.begin()->first;
-	    nrecHit_vs_nsimHit_sim2rec[w]->Fill( assoctrack->numberOfValidHits(),nSimHits);
+
+	int assoc_recoTrack_NValidHits = 0;
+	if(simRecColl.find(tpr) != simRecColl.end()) {
+	  auto const & rt = simRecColl[tpr];
+	  if (!rt.empty()) {
+	    RefToBase<Track> assoc_recoTrack = rt.begin()->first;
+	    edm::LogVerbatim("MuonTrackValidator")<<"-----------------------------associated Track #"<<assoc_recoTrack.key();
+	    TP_is_matched = true;
+	    ats++;
+	    if (assoc_recoTrack->charge() != tpr->charge()) isChargeOK = false;
+	    quality = rt.begin()->second;
+	    assoc_recoTrack_NValidHits = assoc_recoTrack->numberOfValidHits();
+	    edm::LogVerbatim("MuonTrackValidator") << "TrackingParticle #" <<tpr.key()
+						   << " with pt=" << sqrt(momentumTP.perp2())
+						   << " associated with quality:" << quality <<"\n";
 	  }
+	} else {
+	  edm::LogVerbatim("MuonTrackValidator")
+	    << "TrackingParticle #" << tpr.key()
+	    << " with pt,eta,phi: "
+	    << sqrt(momentumTP.perp2()) << " , "<< momentumTP.eta() << " , "<< momentumTP.phi() << " , "
+	    << " NOT associated to any reco::Track" << "\n";
+	}
+
+	// histos for efficiency vs eta
+
+	fillPlotNoFlow(h_simuleta[w], xTPeta);
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assoceta[w], xTPeta);
+	  if (!isChargeOK) fillPlotNoFlow(h_misideta[w], xTPeta);
+	}
+	  
+	// histos for efficiency vs phi	
+	fillPlotNoFlow(h_simulphi[w], TPphi);
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assocphi[w], TPphi);
+	  if (!isChargeOK) fillPlotNoFlow(h_misidphi[w], TPphi);
+	}
+	
+	// histos for efficiency vs pT
+	fillPlotNoFlow(h_simulpT[w], xTPpt);
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assocpT[w], xTPpt);
+	  if (!isChargeOK) fillPlotNoFlow(h_misidpT[w], xTPpt);
+	}
+	
+	// histos for efficiency vs dxy
+	fillPlotNoFlow(h_simuldxy[w], dxySim);
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assocdxy[w], dxySim);
+	  if (!isChargeOK) fillPlotNoFlow(h_misiddxy[w], dxySim);
+	}
+
+	// histos for efficiency vs dz
+	fillPlotNoFlow(h_simuldz[w], dzSim);
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assocdz[w], dzSim);
+	  if (!isChargeOK) fillPlotNoFlow(h_misiddz[w], dzSim);
+	}
+
+	// histos for efficiency vs Radius
+	fillPlotNoFlow(h_simulRpos[w], TPrpos);
+	if (TP_is_matched) fillPlotNoFlow(h_assocRpos[w], TPrpos);
+	
+	// histos for efficiency vs z position
+	fillPlotNoFlow(h_simulZpos[w], TPzpos);
+	if (TP_is_matched) fillPlotNoFlow(h_assocZpos[w], TPzpos);
+
+	// histos for efficiency vs Number of Hits
+	fillPlotNoFlow(h_simulhit[w], nSimHits );
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assochit[w], nSimHits );
+	  nRecHits_vs_nSimHits[w]->Fill(nSimHits, assoc_recoTrack_NValidHits);
+
+	  // charge misid is more useful w.r.t. nRecHits (filled after)
+	  //if (!isChargeOK) fillPlotNoFlow(h_misidhit[w], nSimHits);
+	}
+	
+	// histos for efficiency vs PileUp
+	fillPlotNoFlow(h_simulpu[w], PU_NumInteractions);
+	if (TP_is_matched) {
+	  fillPlotNoFlow(h_assocpu[w], PU_NumInteractions);
+	  if (!isChargeOK) fillPlotNoFlow(h_misidpu[w], PU_NumInteractions);
+	}
+	
       } // End for (TrackingParticleCollection::size_type i=0; i<tPCeff.size(); i++){
-      if (st!=0) h_tracksSIM[w]->Fill(st);
       
       //
       //fill reconstructed track histograms
@@ -557,10 +494,12 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 					     << label[www].label()<<":"
 					     << label[www].instance()
 					     << ": " << trackCollectionSize << "\n";
+
       int at = 0;
       int rT = 0;
-      for(edm::View<Track>::size_type i=0; i<trackCollectionSize; ++i){
+      for(edm::View<Track>::size_type i=0; i<trackCollectionSize; ++i) {
         bool Track_is_matched = false;
+        bool isChargeOK = true;
 	RefToBase<Track> track(trackCollection, i);
 	rT++;
 	
@@ -568,7 +507,7 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 	TrackingParticleRef tpr;
 
 	// new logic (bidirectional)
-	if (BiDirectional_RecoToSim_association) {	
+	if (BiDirectional_RecoToSim_association) {
 	  edm::LogVerbatim("MuonTrackValidator")<<"----------------------------------------Track #"<< track.key();
 	  
 	  if(recSimColl.find(track) != recSimColl.end()) {
@@ -577,18 +516,17 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 	      tpr = tp.begin()->first;	
 	      // RtS and StR must associate the same pair !
 	      if(simRecColl.find(tpr) != simRecColl.end()) {
-		std::vector<std::pair<RefToBase<Track>, double> > track_checkback = simRecColl[tpr];
-		RefToBase<Track> assoc_track_checkback;
-		assoc_track_checkback = track_checkback.begin()->first;
+		auto const & assoc_track_checkback = simRecColl[tpr].begin()->first;
 		
 		if ( assoc_track_checkback.key() == track.key() ) {
 		  edm::LogVerbatim("MuonTrackValidator")<<"------------------associated TrackingParticle #"<<tpr.key();
 		  Track_is_matched = true;
 		  at++;
+		  if (track->charge() != tpr->charge()) isChargeOK = false;
 		  double Purity = tp.begin()->second;
-		  double Quality = track_checkback.begin()->second;
+		  double Quality = simRecColl[tpr].begin()->second;
 		  edm::LogVerbatim("MuonTrackValidator") << "reco::Track #" << track.key() << " with pt=" << track->pt()
-							 << " associated with quality:" << Purity <<"\n";
+							 << " associated with purity:" << Purity <<"\n";
 		  if (MABH) h_PurityVsQuality[w]->Fill(Quality,Purity);
 		}
 	      }
@@ -599,14 +537,15 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 	    edm::LogVerbatim("MuonTrackValidator")
 	      << "reco::Track #" << track.key() << " with pt=" << track->pt() << " NOT associated to any TrackingParticle" << "\n";
 	}
-	// old logic (bugged for collision scenario, still valid for cosmics 2 legs reco)
+	// old logic, valid for cosmics 2-legs reco (bugged for collision scenario)
 	else {
 	  if(recSimColl.find(track) != recSimColl.end()){
 	    tp = recSimColl[track];
 	    if (!tp.empty()) {
-	      Track_is_matched = true;
 	      tpr = tp.begin()->first;
+	      Track_is_matched = true;
 	      at++;
+	      if (track->charge() != tpr->charge()) isChargeOK = false;
 	      edm::LogVerbatim("MuonTrackValidator") << "reco::Track #" << track.key() << " with pt=" << track->pt()
 						     << " associated with quality:" << tp.begin()->second <<"\n";
 	    }
@@ -615,64 +554,74 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 						   << " NOT associated to any TrackingParticle" << "\n";	
 	  }
 	}
-	
-	//Compute fake rate vs eta
-	for (unsigned int f=0; f<etaintervals[w].size()-1; f++){
-	  if (getEta(track->momentum().eta())>etaintervals[w][f]&&
-	      getEta(track->momentum().eta())<etaintervals[w][f+1]) {
-	    totRECeta[w][f]++;
-	    if (Track_is_matched) {
-	      totASS2eta[w][f]++;
-	    }
-	  }
-	} // End for (unsigned int f=0; f<etaintervals[w].size()-1; f++){
 
-	for (unsigned int f=0; f<phiintervals[w].size()-1; f++){
-	  if (track->momentum().phi()>phiintervals[w][f]&&
-	      track->momentum().phi()<phiintervals[w][f+1]) {
-	    totREC_phi[w][f]++;
-	    if (Track_is_matched) {
-	      totASS2_phi[w][f]++;
-	    }
-	  }
-	} // End for (unsigned int f=0; f<phiintervals[w].size()-1; f++){
+	// 
+	int nRecHits = track->numberOfValidHits();
+	edm::LogVerbatim("MuonTrackValidator") << "\t N valid rechits = "<< nRecHits <<"\n";
 	
-	for (unsigned int f=0; f<pTintervals[w].size()-1; f++){
-	  if (getPt(sqrt(track->momentum().perp2()))>pTintervals[w][f]&&
-	      getPt(sqrt(track->momentum().perp2()))<pTintervals[w][f+1]) {
-	    totRECpT[w][f]++;
-	    if (Track_is_matched) {
-	      totASS2pT[w][f]++;
-	    }
-	  }
-	} // End for (unsigned int f=0; f<pTintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<dxyintervals[w].size()-1; f++){
-	  if (track->dxy(bs.position())>dxyintervals[w][f]&&
-	      track->dxy(bs.position())<dxyintervals[w][f+1]) {
-	    totREC_dxy[w][f]++;
-	    if (Track_is_matched) {
-	      totASS2_dxy[w][f]++;
-	    }
-	  }
-	} // End for (unsigned int f=0; f<dxyintervals[w].size()-1; f++){
-	
-	for (unsigned int f=0; f<dzintervals[w].size()-1; f++){
-	  if (track->dz(bs.position())>dzintervals[w][f]&&
-	      track->dz(bs.position())<dzintervals[w][f+1]) {
-	    totREC_dz[w][f]++;
-	    if (Track_is_matched) {
-	      totASS2_dz[w][f]++;
-	    }
-	  }
-	} // End for (unsigned int f=0; f<dzintervals[w].size()-1; f++){
+	double etaRec = track->eta();
+	double xetaRec = getEta(etaRec);
 
-	int tmp = std::min((int)track->found(),int(maxHit-1));
-	totREC_hit[w][tmp]++;
-	if (Track_is_matched) totASS2_hit[w][tmp]++;
+	double ptRec = track->pt();
+	double xptRec = getPt(ptRec);
+
+	double qoverpRec = track->qoverp();
+	double phiRec = track->phi();
+	double thetaRec = track->theta();
+	double dxyRec = track->dxy(bs.position());
+	double dzRec = track->dz(bs.position()); 
+
+	double qoverpError = track->qoverpError();
+	double ptError = track->ptError();
+        double thetaError = track->thetaError();
+	double phiError = track->phiError();
+	double dxyError = track->dxyError();
+	double dzError = track->dzError();
+
+	// histos for fake rate vs eta
+	fillPlotNoFlow(h_recoeta[w], xetaRec);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2eta[w], xetaRec);
+	}
+
+	// histos for fake rate vs phi
+	fillPlotNoFlow(h_recophi[w], phiRec);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2phi[w], phiRec);
+	}
 	
-	edm::LogVerbatim("MuonTrackValidator") << "\t N valid rechits = "<< (int)track->found() <<"\n";
+	// histos for fake rate vs pT
+	fillPlotNoFlow(h_recopT[w], xptRec);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2pT[w], xptRec);
+	}
 	
+	// histos for fake rate vs dxy
+	fillPlotNoFlow(h_recodxy[w], dxyRec);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2dxy[w], dxyRec);
+	}
+	
+	// histos for fake rate vs dz
+	fillPlotNoFlow(h_recodz[w], dzRec);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2dz[w], dzRec);
+	}
+
+	// histos for fake rate vs Number of RecHits in track 
+	fillPlotNoFlow(h_recohit[w], nRecHits);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2hit[w], nRecHits);
+	  // charge misid w.r.t. nRecHits
+	  if (!isChargeOK) fillPlotNoFlow(h_misidhit[w], nRecHits);
+	}
+
+	// histos for fake rate vs Number of PU interactions
+	fillPlotNoFlow(h_recopu[w], PU_NumInteractions);
+	if (Track_is_matched) {
+	  fillPlotNoFlow(h_assoc2pu[w], PU_NumInteractions);
+	}
+
 	// Fill other histos
 	TrackingParticle* tpp = const_cast<TrackingParticle*>(tpr.get());
 	// TrackingParticle parameters at point of closest approach to the beamline
@@ -705,287 +654,143 @@ void MuonTrackValidator::analyze(const edm::Event& event, const edm::EventSetup&
 	else if (associators[ww]=="trackAssociatorByHits"){
 	  double fraction = tp.begin()->second;
 	  h_assocFraction[www]->Fill(fraction);
-	  h_assocSharedHit[www]->Fill(fraction*track->numberOfValidHits());
+	  h_assocSharedHit[www]->Fill(fraction*nRecHits);
 	}
 	
-	//nchi2 and hits global distributions
+	h_charge[w]->Fill(track->charge());
+
+	// Hits
+	h_nhits[w]->Fill(nRecHits);
+	nhits_vs_eta[w]->Fill(xetaRec, nRecHits);
+	nhits_vs_phi[w]->Fill(phiRec, nRecHits);
+
+	if (do_MUOhitsPlots) {
+	  nDThits_vs_eta[w]->Fill(xetaRec,track->hitPattern().numberOfValidMuonDTHits());
+	  nCSChits_vs_eta[w]->Fill(xetaRec,track->hitPattern().numberOfValidMuonCSCHits());
+	  nRPChits_vs_eta[w]->Fill(xetaRec,track->hitPattern().numberOfValidMuonRPCHits());	
+	  if(useGEMs_) nGEMhits_vs_eta[w]->Fill(xetaRec,track->hitPattern().numberOfValidMuonGEMHits());
+	  if(useME0_) nME0hits_vs_eta[w]->Fill(xetaRec,track->hitPattern().numberOfValidMuonME0Hits());
+	}
+
+	if (do_TRKhitsPlots) {
+	  nTRK_LayersWithMeas_vs_eta[w]->Fill(xetaRec,track->hitPattern().trackerLayersWithMeasurement());
+	  nPixel_LayersWithMeas_vs_eta[w]->Fill(xetaRec,track->hitPattern().pixelLayersWithMeasurement());
+	  h_nlosthits[w]->Fill(track->numberOfLostHits());
+	  h_nmisslayers_inner[w]->Fill(track->hitPattern().numberOfLostHits(reco::HitPattern::MISSING_INNER_HITS));
+	  h_nmisslayers_outer[w]->Fill(track->hitPattern().numberOfLostHits(reco::HitPattern::MISSING_OUTER_HITS));
+	  nlosthits_vs_eta[w]->Fill(xetaRec,track->numberOfLostHits());
+	}
+
+	// normalized chi2
 	h_nchi2[w]->Fill(track->normalizedChi2());
 	h_nchi2_prob[w]->Fill(TMath::Prob(track->chi2(),(int)track->ndof()));
-	h_hits[w]->Fill(track->numberOfValidHits());
-	h_losthits[w]->Fill(track->numberOfLostHits());
-	chi2_vs_nhits[w]->Fill(track->numberOfValidHits(),track->normalizedChi2());
-	h_charge[w]->Fill( track->charge() );
+	chi2_vs_nhits[w]->Fill(nRecHits, track->normalizedChi2());
+	chi2_vs_eta[w]->Fill(xetaRec, track->normalizedChi2());
+	chi2_vs_phi[w]->Fill(phiRec,track->normalizedChi2());
 	
 	double ptSim = sqrt(momentumTP.perp2());
-	double qoverpSim = tpr->charge()/sqrt(momentumTP.x()*momentumTP.x()+momentumTP.y()*momentumTP.y()+momentumTP.z()*momentumTP.z());
+	double xptSim = getPt(ptSim);
+	double qoverpSim = tpr->charge() /
+	  sqrt(momentumTP.x()*momentumTP.x()+momentumTP.y()*momentumTP.y()+momentumTP.z()*momentumTP.z());
+	double etaSim = momentumTP.eta();
 	double thetaSim = momentumTP.theta();
-	double lambdaSim = M_PI/2-momentumTP.theta();
 	double phiSim = momentumTP.phi();
 	double dxySim = TrackingParticleIP::dxy(vertexTP, momentumTP, bs.position());
 	double dzSim = TrackingParticleIP::dz(vertexTP, momentumTP, bs.position());
 	
-	// removed unused variable, left this in case it has side effects
-	track->parameters();
-	
-	double qoverpRec(0);
-	double qoverpErrorRec(0);
-	double ptRec(0);
-	double ptErrorRec(0);
-	double lambdaRec(0);
-	double lambdaErrorRec(0);
-	double phiRec(0);
-	double phiErrorRec(0);
-	
-	//loop to decide whether to take gsfTrack (utilisation of mode-function) or common track
-	const GsfTrack* gsfTrack(nullptr);
-	if(useGsf){
-	  gsfTrack = dynamic_cast<const GsfTrack*>(&(*track));
-	  if (gsfTrack==nullptr) edm::LogInfo("MuonTrackValidator") << "Trying to access mode for a non-GsfTrack";
-	}
-	
-	if (gsfTrack) {
-	  // get values from mode
-	  getRecoMomentum(*gsfTrack, ptRec, ptErrorRec, qoverpRec, qoverpErrorRec,
-			  lambdaRec,lambdaErrorRec, phiRec, phiErrorRec);
-	}
-	
-	else {
-	  // get values from track (without mode)
-	  getRecoMomentum(*track, ptRec, ptErrorRec, qoverpRec, qoverpErrorRec,
-			  lambdaRec,lambdaErrorRec, phiRec, phiErrorRec);
-	}
-	
-	double thetaRec = track->theta();
-	double ptError = ptErrorRec;
-	double ptres = ptRec - ptSim;
-	double etares = track->eta()-momentumTP.Eta();
-	double dxyRec = track->dxy(bs.position());
-	double dzRec = track->dz(bs.position());
-	// eta residue; pt, k, theta, phi, dxy, dz pulls
-	double qoverpPull=(qoverpRec-qoverpSim)/qoverpErrorRec;
-	double thetaPull=(lambdaRec-lambdaSim)/lambdaErrorRec;
+	double etares = etaRec - etaSim;
+	double ptRelRes = (ptRec - ptSim) / ptSim;    // relative residual -> resolution
+	double ptPull = (ptRec - ptSim) / ptError; 
+	double qoverpPull = (qoverpRec-qoverpSim) / qoverpError;
+	double thetaPull = (thetaRec - thetaSim) / thetaError;
 	double phiDiff = phiRec - phiSim;
 	if (abs(phiDiff) > M_PI) {
 	  if (phiDiff >0.) phiDiff = phiDiff - 2.*M_PI;
 	  else phiDiff = phiDiff + 2.*M_PI;
 	}
-	double phiPull=phiDiff/phiErrorRec;
-	double dxyPull=(dxyRec-dxySim)/track->dxyError();
-	double dzPull=(dzRec-dzSim)/track->dzError();
+	double phiPull = phiDiff / phiError;
+	double dxyPull = (dxyRec-dxySim) / dxyError;
+	double dzPull = (dzRec-dzSim) / dzError;
 	
-	double contrib_Qoverp = ((qoverpRec-qoverpSim)/qoverpErrorRec)*((qoverpRec-qoverpSim)/qoverpErrorRec)/5;
-	double contrib_dxy = ((dxyRec-dxySim)/track->dxyError())*((dxyRec-dxySim)/track->dxyError())/5;
-	double contrib_dz = ((dzRec-dzSim)/track->dzError())*((dzRec-dzSim)/track->dzError())/5;
-	double contrib_theta = ((lambdaRec-lambdaSim)/lambdaErrorRec)*((lambdaRec-lambdaSim)/lambdaErrorRec)/5;
-	double contrib_phi = (phiDiff/phiErrorRec)*(phiDiff/phiErrorRec)/5;
+	h_etaRes[w]->Fill(etares);
+	etares_vs_eta[w]->Fill(xetaRec, etares);
 	
-	edm::LogVerbatim("MuonTrackValidator") << "assocChi2=" << tp.begin()->second << "\n"
-					       << "" << "\n"
-					       << "ptREC=" << ptRec << "\n"
-					       << "etaREC=" << track->eta() << "\n"
-					       << "qoverpREC=" << qoverpRec << "\n"
-					       << "dxyREC=" << dxyRec << "\n"
-					       << "dzREC=" << dzRec << "\n"
-					       << "thetaREC=" << track->theta() << "\n"
-					       << "phiREC=" << phiRec << "\n"
-					       << "" << "\n"
-					       << "qoverpError()=" << qoverpErrorRec << "\n"
-					       << "dxyError()=" << track->dxyError() << "\n"
-					       << "dzError()=" << track->dzError() << "\n"
-					       << "thetaError()=" << lambdaErrorRec << "\n"
-					       << "phiError()=" << phiErrorRec << "\n"
-					       << "" << "\n"
-					       << "ptSIM=" << ptSim << "\n"
-					       << "etaSIM=" << momentumTP.Eta() << "\n"
-					       << "qoverpSIM=" << qoverpSim << "\n"
-					       << "dxySIM=" << dxySim << "\n"
-					       << "dzSIM=" << dzSim << "\n"
-					       << "thetaSIM=" << M_PI/2-lambdaSim << "\n"
-					       << "phiSIM=" << phiSim << "\n"
-					       << "" << "\n"
-					       << "contrib_Qoverp=" << contrib_Qoverp << "\n"
-					       << "contrib_dxy=" << contrib_dxy << "\n"
-					       << "contrib_dz=" << contrib_dz << "\n"
-					       << "contrib_theta=" << contrib_theta << "\n"
-					       << "contrib_phi=" << contrib_phi << "\n"
-					       << "" << "\n"
-					       <<"chi2PULL="<<contrib_Qoverp+contrib_dxy+contrib_dz+contrib_theta+contrib_phi<<"\n";
-	
-	h_pullQoverp[w]->Fill(qoverpPull);
-	h_pullTheta[w]->Fill(thetaPull);
-	h_pullPhi[w]->Fill(phiPull);
-	h_pullDxy[w]->Fill(dxyPull);
-	h_pullDz[w]->Fill(dzPull);
-	
-	h_pt[w]->Fill(ptres/ptError);
-	h_eta[w]->Fill(etares);
-	etares_vs_eta[w]->Fill(getEta(track->eta()),etares);
-	
-	//chi2 and #hit vs eta: fill 2D histos
-	chi2_vs_eta[w]->Fill(getEta(track->eta()),track->normalizedChi2());
-	nhits_vs_eta[w]->Fill(getEta(track->eta()),track->numberOfValidHits());
-	nDThits_vs_eta[w]->Fill(getEta(track->eta()),track->hitPattern().numberOfValidMuonDTHits());
-	nCSChits_vs_eta[w]->Fill(getEta(track->eta()),track->hitPattern().numberOfValidMuonCSCHits());
-	nRPChits_vs_eta[w]->Fill(getEta(track->eta()),track->hitPattern().numberOfValidMuonRPCHits());
-    if(useGEMs_) nGEMhits_vs_eta[w]->Fill(getEta(track->eta()),track->hitPattern().numberOfValidMuonGEMHits());
-    if(useME0_) nME0hits_vs_eta[w]->Fill(getEta(track->eta()),track->hitPattern().numberOfValidMuonME0Hits());
-	nlosthits_vs_eta[w]->Fill(getEta(track->eta()),track->numberOfLostHits());
-	
-	//resolution of track params: fill 2D histos
-	dxyres_vs_eta[w]->Fill(getEta(track->eta()),dxyRec-dxySim);
-	ptres_vs_eta[w]->Fill(getEta(track->eta()),(ptRec-ptSim)/ptRec);
-	dzres_vs_eta[w]->Fill(getEta(track->eta()),dzRec-dzSim);
-	phires_vs_eta[w]->Fill(getEta(track->eta()),phiDiff);
-	cotThetares_vs_eta[w]->Fill(getEta(track->eta()), cos(thetaRec)/sin(thetaRec) - cos(thetaSim)/sin(thetaSim));
-	
-	//same as before but vs pT
-	dxyres_vs_pt[w]->Fill(getPt(ptRec),dxyRec-dxySim);
-	ptres_vs_pt[w]->Fill(getPt(ptRec),(ptRec-ptSim)/ptRec);
-	dzres_vs_pt[w]->Fill(getPt(ptRec),dzRec-dzSim);
-	phires_vs_pt[w]->Fill(getPt(ptRec),phiDiff);
-	cotThetares_vs_pt[w]->Fill(getPt(ptRec), cos(thetaRec)/sin(thetaRec) - cos(thetaSim)/sin(thetaSim));
-	
-	//pulls of track params vs eta: fill 2D histos
-	dxypull_vs_eta[w]->Fill(getEta(track->eta()),dxyPull);
-	ptpull_vs_eta[w]->Fill(getEta(track->eta()),ptres/ptError);
-	dzpull_vs_eta[w]->Fill(getEta(track->eta()),dzPull);
-	phipull_vs_eta[w]->Fill(getEta(track->eta()),phiPull);
-	thetapull_vs_eta[w]->Fill(getEta(track->eta()),thetaPull);
-	
-	//plots vs phi
-	nhits_vs_phi[w]->Fill(phiRec,track->numberOfValidHits());
-	chi2_vs_phi[w]->Fill(phiRec,track->normalizedChi2());
-	ptmean_vs_eta_phi[w]->Fill(phiRec,getEta(track->eta()),ptRec);
-	phimean_vs_eta_phi[w]->Fill(phiRec,getEta(track->eta()),phiRec);
-	ptres_vs_phi[w]->Fill(phiRec,(ptRec-ptSim)/ptRec);
-	phires_vs_phi[w]->Fill(phiRec,phiDiff);
-	ptpull_vs_phi[w]->Fill(phiRec,ptres/ptError);
-	phipull_vs_phi[w]->Fill(phiRec,phiPull);
+	ptres_vs_eta[w]->Fill(xetaRec,ptRelRes);
+	ptres_vs_pt[w]->Fill(xptSim,ptRelRes);
+	ptres_vs_phi[w]->Fill(phiRec,ptRelRes);
+	h_ptpull[w]->Fill(ptPull);
+	ptpull_vs_eta[w]->Fill(xetaRec,ptPull);
+	ptpull_vs_phi[w]->Fill(phiRec,ptPull);
+	h_qoverppull[w]->Fill(qoverpPull);
+
+	thetaCotres_vs_eta[w]->Fill(xetaRec, cos(thetaRec)/sin(thetaRec) - cos(thetaSim)/sin(thetaSim));
+	thetaCotres_vs_pt[w]->Fill(xptSim, cos(thetaRec)/sin(thetaRec) - cos(thetaSim)/sin(thetaSim));
+	h_thetapull[w]->Fill(thetaPull);
+	thetapull_vs_eta[w]->Fill(xetaRec,thetaPull);
 	thetapull_vs_phi[w]->Fill(phiRec,thetaPull);
+
+	phires_vs_eta[w]->Fill(xetaRec,phiDiff);
+	phires_vs_pt[w]->Fill(xptSim,phiDiff);
+	phires_vs_phi[w]->Fill(phiRec,phiDiff);
+	h_phipull[w]->Fill(phiPull);
+	phipull_vs_eta[w]->Fill(xetaRec,phiPull);
+	phipull_vs_phi[w]->Fill(phiRec,phiPull);
+
+	dxyres_vs_eta[w]->Fill(xetaRec,dxyRec-dxySim);
+	dxyres_vs_pt[w]->Fill(xptSim,dxyRec-dxySim);
+	h_dxypull[w]->Fill(dxyPull);
+	dxypull_vs_eta[w]->Fill(xetaRec,dxyPull);
+
+	dzres_vs_eta[w]->Fill(xetaRec,dzRec-dzSim);
+	dzres_vs_pt[w]->Fill(xptSim,dzRec-dzSim);
+	h_dzpull[w]->Fill(dzPull);	
+	dzpull_vs_eta[w]->Fill(xetaRec,dzPull);
+
+	double contrib_Qoverp = qoverpPull*qoverpPull/5;
+	double contrib_dxy = dxyPull*dxyPull/5;
+	double contrib_dz = dzPull*dzPull/5;
+	double contrib_theta = thetaPull*thetaPull/5;
+	double contrib_phi = phiPull*phiPull/5;
+	double assoChi2 = contrib_Qoverp+contrib_dxy+contrib_dz+contrib_theta+contrib_phi;
+
+	edm::LogVerbatim("MuonTrackValidator") << "normalized Chi2 (track 5-dofs matching) = "<<assoChi2 <<"\n"
+					       << "\t contrib_Qoverp = " << contrib_Qoverp << "\n"
+					       << "\t contrib_theta = " << contrib_theta << "\n"
+					       << "\t contrib_phi = " << contrib_phi << "\n"
+					       << "\t contrib_dxy = " << contrib_dxy << "\n"
+					       << "\t contrib_dz = " << contrib_dz << "\n";
 	
-	int nSimHits = 0;
-	if (usetracker && usemuon) {
-	  nSimHits= tpr.get()->numberOfHits();
-	}
-	else if (!usetracker && usemuon) {
-	  nSimHits= tpr.get()->numberOfHits() - tpr.get()->numberOfTrackerHits();
-	}
-	else if (usetracker && !usemuon) {
-	  nSimHits=tpr.get()->numberOfTrackerHits();
-	}
-	
-	nrecHit_vs_nsimHit_rec2sim[w]->Fill(track->numberOfValidHits(), nSimHits);
-	
-      } // End of for(View<Track>::size_type i=0; i<trackCollectionSize; ++i){
+	edm::LogVerbatim("MuonTrackValidator") << "ptRec = " << ptRec << "\n"
+					       << "etaRec = " << etaRec << "\n"
+					       << "qoverpRec = " << qoverpRec << "\n"
+					       << "thetaRec = " << thetaRec << "\n"
+					       << "phiRec = " << phiRec << "\n"
+					       << "dxyRec = " << dxyRec << "\n"
+					       << "dzRec = " << dzRec << "\n"
+					       << "" << "\n"
+					       << "qoverpError = " << qoverpError << "\n"
+					       << "thetaError = " << thetaError << "\n"
+					       << "phiError = " << phiError << "\n"
+					       << "dxyError = " << dxyError << "\n"
+					       << "dzError = " << dzError << "\n"
+					       << "" << "\n"
+					       << "ptSim = " << ptSim << "\n"
+					       << "etaSim = " << etaSim << "\n"
+					       << "qoverpSim = " << qoverpSim << "\n"
+					       << "thetaSim = " << thetaSim << "\n"
+					       << "phiSim = " << phiSim << "\n"
+					       << "dxySim = " << dxySim << "\n"
+					       << "dzSim = " << dzSim << "\n";
+      } // End of for(edm::View<Track>::size_type i=0; i<trackCollectionSize; ++i) {
       
-      if (at!=0) h_tracks[w]->Fill(at);
+      h_tracks[w]->Fill(at);
       h_fakes[w]->Fill(rT-at);
       edm::LogVerbatim("MuonTrackValidator") << "Total Simulated: " << st << "\n"
 					     << "Total Associated (simToReco): " << ats << "\n"
 					     << "Total Reconstructed: " << rT << "\n"
 					     << "Total Associated (recoToSim): " << at << "\n"
 					     << "Total Fakes: " << rT-at << "\n";
-      nrec_vs_nsim[w]->Fill(rT,st);
-      w++;
+       w++;
     } // End of for (unsigned int www=0;www<label.size();www++){
   } //END of for (unsigned int ww=0;ww<associators.size();ww++){
-}
-
-void MuonTrackValidator::endRun(Run const&, EventSetup const&) {
-  
-  int w=0;
-  for (unsigned int ww=0;ww<associators.size();ww++){
-    for (unsigned int www=0;www<label.size();www++){
-      
-      //chi2 and #hit vs eta: get mean from 2D histos
-      doProfileX(chi2_vs_eta[w],h_chi2meanh[w]);
-      doProfileX(nhits_vs_eta[w],h_hits_eta[w]);
-      doProfileX(nDThits_vs_eta[w],h_DThits_eta[w]);
-      doProfileX(nCSChits_vs_eta[w],h_CSChits_eta[w]);
-      doProfileX(nRPChits_vs_eta[w],h_RPChits_eta[w]);
-      if (useGEMs_) doProfileX(nGEMhits_vs_eta[w],h_GEMhits_eta[w]);
-      if (useME0_) doProfileX(nME0hits_vs_eta[w],h_ME0hits_eta[w]);
-      
-      doProfileX(nlosthits_vs_eta[w],h_losthits_eta[w]);
-      //vs phi
-      doProfileX(chi2_vs_nhits[w],h_chi2meanhitsh[w]);
-      doProfileX(chi2_vs_phi[w],h_chi2mean_vs_phi[w]);
-      doProfileX(nhits_vs_phi[w],h_hits_phi[w]);
-
-      fillPlotFromVector(h_recoeta[w],totRECeta[w]);
-      fillPlotFromVector(h_simuleta[w],totSIMeta[w]);
-      fillPlotFromVector(h_assoceta[w],totASSeta[w]);
-      fillPlotFromVector(h_assoc2eta[w],totASS2eta[w]);
-
-      fillPlotFromVector(h_recopT[w],totRECpT[w]);
-      fillPlotFromVector(h_simulpT[w],totSIMpT[w]);
-      fillPlotFromVector(h_assocpT[w],totASSpT[w]);
-      fillPlotFromVector(h_assoc2pT[w],totASS2pT[w]);
-
-      fillPlotFromVector(h_recohit[w],totREC_hit[w]);
-      fillPlotFromVector(h_simulhit[w],totSIM_hit[w]);
-      fillPlotFromVector(h_assochit[w],totASS_hit[w]);
-      fillPlotFromVector(h_assoc2hit[w],totASS2_hit[w]);
-
-      fillPlotFromVector(h_recophi[w],totREC_phi[w]);
-      fillPlotFromVector(h_simulphi[w],totSIM_phi[w]);
-      fillPlotFromVector(h_assocphi[w],totASS_phi[w]);
-      fillPlotFromVector(h_assoc2phi[w],totASS2_phi[w]);
-
-      fillPlotFromVector(h_recodxy[w],totREC_dxy[w]);
-      fillPlotFromVector(h_simuldxy[w],totSIM_dxy[w]);
-      fillPlotFromVector(h_assocdxy[w],totASS_dxy[w]);
-      fillPlotFromVector(h_assoc2dxy[w],totASS2_dxy[w]);
-
-      fillPlotFromVector(h_recodz[w],totREC_dz[w]);
-      fillPlotFromVector(h_simuldz[w],totSIM_dz[w]);
-      fillPlotFromVector(h_assocdz[w],totASS_dz[w]);
-      fillPlotFromVector(h_assoc2dz[w],totASS2_dz[w]);
-
-      fillPlotFromVector(h_simulvertpos[w],totSIM_vertpos[w]);
-      fillPlotFromVector(h_assocvertpos[w],totASS_vertpos[w]);
-
-      fillPlotFromVector(h_simulzpos[w],totSIM_zpos[w]);
-      fillPlotFromVector(h_assoczpos[w],totASS_zpos[w]);
-      
-      if (MABH) {
-	fillPlotFromVector(h_assoceta_Quality05[w] ,totASSeta_Quality05[w]);
-	fillPlotFromVector(h_assoceta_Quality075[w],totASSeta_Quality075[w]);
-	fillPlotFromVector(h_assocpT_Quality05[w] ,totASSpT_Quality05[w]);
-	fillPlotFromVector(h_assocpT_Quality075[w],totASSpT_Quality075[w]);
-	fillPlotFromVector(h_assocphi_Quality05[w] ,totASS_phi_Quality05[w]);
-	fillPlotFromVector(h_assocphi_Quality075[w],totASS_phi_Quality075[w]);
-      }
-      
-      w++;
-    }
-  }
-  
-  if ( !out.empty() && dbe_ ) dbe_->save(out);
-}
-
-void MuonTrackValidator::getRecoMomentum (const reco::Track& track,
-					  double& pt, double& ptError, double& qoverp, double& qoverpError,
-					  double& lambda,double& lambdaError, double& phi, double& phiError ) const {
-  pt = track.pt();
-  ptError = track.ptError();
-  qoverp = track.qoverp();
-  qoverpError = track.qoverpError();
-  lambda = track.lambda();
-  lambdaError = track.lambdaError();
-  phi = track.phi();
-  phiError = track.phiError();
-}
-
-void MuonTrackValidator::getRecoMomentum (const reco::GsfTrack& gsfTrack,
-					  double& pt, double& ptError, double& qoverp, double& qoverpError,
-					  double& lambda,double& lambdaError, double& phi, double& phiError ) const {
-  pt = gsfTrack.ptMode();
-  ptError = gsfTrack.ptModeError();
-  qoverp = gsfTrack.qoverpMode();
-  qoverpError = gsfTrack.qoverpModeError();
-  lambda = gsfTrack.lambdaMode();
-  lambdaError = gsfTrack.lambdaModeError();
-  phi = gsfTrack.phiMode();
-  phiError = gsfTrack.phiModeError(); 
 }
