+++ Validation/RecoMuon/plugins/MuonTrackValidatorBase.h
@@ -19,7 +19,7 @@
 #include "DataFormats/BeamSpot/interface/BeamSpot.h"
 #include "SimDataFormats/TrackingAnalysis/interface/TrackingParticle.h"
 #include "SimDataFormats/TrackingAnalysis/interface/TrackingParticleFwd.h"
-
+#include "SimDataFormats/PileupSummaryInfo/interface/PileupSummaryInfo.h"
 
 #include "DQMServices/Core/interface/MonitorElement.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
@@ -33,6 +33,7 @@
 #include <TH1F.h>
 #include <TH2F.h>
 
+
 class DQMStore;
 class MuonTrackValidatorBase {
  public:
@@ -42,6 +43,7 @@ class MuonTrackValidatorBase {
       bsSrc_Token = iC.consumes<reco::BeamSpot>(bsSrc);
       tp_effic_Token = iC.consumes<TrackingParticleCollection>(label_tp_effic);
       tp_fake_Token = iC.consumes<TrackingParticleCollection>(label_tp_fake);
+      pileupinfo_Token = iC.consumes<std::vector<PileupSummaryInfo> >(label_pileupinfo);
       for (unsigned int www=0;www<label.size();www++){
 	track_Collection_Token[www] = iC.consumes<edm::View<reco::Track> >(label[www]);
       }
@@ -49,71 +51,105 @@ class MuonTrackValidatorBase {
 
   MuonTrackValidatorBase(const edm::ParameterSet& pset):
     label(pset.getParameter< std::vector<edm::InputTag> >("label")),
-    usetracker(pset.getParameter<bool>("usetracker")),
-    usemuon(pset.getParameter<bool>("usemuon")),
     bsSrc(pset.getParameter< edm::InputTag >("beamSpot")),
     label_tp_effic(pset.getParameter< edm::InputTag >("label_tp_effic")),
     label_tp_fake(pset.getParameter< edm::InputTag >("label_tp_fake")),
+    label_pileupinfo(pset.getParameter< edm::InputTag >("label_pileupinfo")),
     associators(pset.getParameter< std::vector<std::string> >("associators")),
     out(pset.getParameter<std::string>("outputFile")),
     parametersDefiner(pset.getParameter<std::string>("parametersDefiner")),
-    min(pset.getParameter<double>("min")),
-    max(pset.getParameter<double>("max")),
-    nint(pset.getParameter<int>("nint")),
-    useFabs(pset.getParameter<bool>("useFabsEta")),
-    minpT(pset.getParameter<double>("minpT")),
-    maxpT(pset.getParameter<double>("maxpT")),
-    nintpT(pset.getParameter<int>("nintpT")),
-    minHit(pset.getParameter<double>("minHit")),
-    maxHit(pset.getParameter<double>("maxHit")),
-    nintHit(pset.getParameter<int>("nintHit")),
-    minPhi(pset.getParameter<double>("minPhi")),
-    maxPhi(pset.getParameter<double>("maxPhi")),
-    nintPhi(pset.getParameter<int>("nintPhi")),
-    minDxy(pset.getParameter<double>("minDxy")),
-    maxDxy(pset.getParameter<double>("maxDxy")),
-    nintDxy(pset.getParameter<int>("nintDxy")),
-    minDz(pset.getParameter<double>("minDz")),
-    maxDz(pset.getParameter<double>("maxDz")),
-    nintDz(pset.getParameter<int>("nintDz")),
-    minVertpos(pset.getParameter<double>("minVertpos")),
-    maxVertpos(pset.getParameter<double>("maxVertpos")),
-    nintVertpos(pset.getParameter<int>("nintVertpos")),
-    minZpos(pset.getParameter<double>("minZpos")),
-    maxZpos(pset.getParameter<double>("maxZpos")),
-    nintZpos(pset.getParameter<int>("nintZpos")),
-    useInvPt(pset.getParameter<bool>("useInvPt")),
-    //
-    ptRes_rangeMin(pset.getParameter<double>("ptRes_rangeMin")),
-    ptRes_rangeMax(pset.getParameter<double>("ptRes_rangeMax")),
-    phiRes_rangeMin(pset.getParameter<double>("phiRes_rangeMin")),
-    phiRes_rangeMax(pset.getParameter<double>("phiRes_rangeMax")),
-    cotThetaRes_rangeMin(pset.getParameter<double>("cotThetaRes_rangeMin")),
-    cotThetaRes_rangeMax(pset.getParameter<double>("cotThetaRes_rangeMax")),
-    dxyRes_rangeMin(pset.getParameter<double>("dxyRes_rangeMin")),
-    dxyRes_rangeMax(pset.getParameter<double>("dxyRes_rangeMax")),
-    dzRes_rangeMin(pset.getParameter<double>("dzRes_rangeMin")),
-    dzRes_rangeMax(pset.getParameter<double>("dzRes_rangeMax")),
-    ptRes_nbin(pset.getParameter<int>("ptRes_nbin")),
-    cotThetaRes_nbin(pset.getParameter<int>("cotThetaRes_nbin")),
-    phiRes_nbin(pset.getParameter<int>("phiRes_nbin")),
-    dxyRes_nbin(pset.getParameter<int>("dxyRes_nbin")),
-    dzRes_nbin(pset.getParameter<int>("dzRes_nbin")),
-    ignoremissingtkcollection_(pset.getUntrackedParameter<bool>("ignoremissingtrackcollection",false)),
-    useLogPt(pset.getUntrackedParameter<bool>("useLogPt",false))
-    //
+    muonHistoParameters(pset.getParameter<edm::ParameterSet>("muonHistoParameters")),
+    ignoremissingtkcollection_(pset.getUntrackedParameter<bool>("ignoremissingtrackcollection",false))
+
     {
-      dbe_ = edm::Service<DQMStore>().operator->();
-      if(useLogPt){
-	maxpT=log10(maxpT);
-	minpT=log10(minpT);
+      minEta = muonHistoParameters.getParameter<double>("minEta");
+      maxEta = muonHistoParameters.getParameter<double>("maxEta");
+      nintEta = muonHistoParameters.getParameter<int>("nintEta");
+      useFabsEta = muonHistoParameters.getParameter<bool>("useFabsEta");
+      minPt = muonHistoParameters.getParameter<double>("minPt");
+      maxPt = muonHistoParameters.getParameter<double>("maxPt");
+      nintPt = muonHistoParameters.getParameter<int>("nintPt");
+      useLogPt = muonHistoParameters.getUntrackedParameter<bool>("useLogPt",false);
+      useInvPt = muonHistoParameters.getParameter<bool>("useInvPt");
+      minNHit = muonHistoParameters.getParameter<double>("minNHit");
+      maxNHit = muonHistoParameters.getParameter<double>("maxNHit");
+      nintNHit = muonHistoParameters.getParameter<int>("nintNHit");
+      //
+      minDTHit = muonHistoParameters.getParameter<double>("minDTHit");
+      maxDTHit = muonHistoParameters.getParameter<double>("maxDTHit");
+      nintDTHit = muonHistoParameters.getParameter<int>("nintDTHit");
+      //
+      minCSCHit = muonHistoParameters.getParameter<double>("minCSCHit");
+      maxCSCHit = muonHistoParameters.getParameter<double>("maxCSCHit");
+      nintCSCHit = muonHistoParameters.getParameter<int>("nintCSCHit");
+      //
+      minRPCHit = muonHistoParameters.getParameter<double>("minRPCHit");
+      maxRPCHit = muonHistoParameters.getParameter<double>("maxRPCHit");
+      nintRPCHit = muonHistoParameters.getParameter<int>("nintRPCHit");
+      //
+      minLayers = muonHistoParameters.getParameter<double>("minLayers");
+      maxLayers = muonHistoParameters.getParameter<double>("maxLayers");
+      nintLayers = muonHistoParameters.getParameter<int>("nintLayers");
+      minPixels = muonHistoParameters.getParameter<double>("minPixels");
+      maxPixels = muonHistoParameters.getParameter<double>("maxPixels");
+      nintPixels = muonHistoParameters.getParameter<int>("nintPixels");
+      minPhi = muonHistoParameters.getParameter<double>("minPhi");
+      maxPhi = muonHistoParameters.getParameter<double>("maxPhi");
+      nintPhi = muonHistoParameters.getParameter<int>("nintPhi");
+      minDxy = muonHistoParameters.getParameter<double>("minDxy");
+      maxDxy = muonHistoParameters.getParameter<double>("maxDxy");
+      nintDxy = muonHistoParameters.getParameter<int>("nintDxy");
+      minDz = muonHistoParameters.getParameter<double>("minDz");
+      maxDz = muonHistoParameters.getParameter<double>("maxDz");
+      nintDz = muonHistoParameters.getParameter<int>("nintDz");
+      minRpos = muonHistoParameters.getParameter<double>("minRpos");
+      maxRpos = muonHistoParameters.getParameter<double>("maxRpos");
+      nintRpos = muonHistoParameters.getParameter<int>("nintRpos");
+      minZpos = muonHistoParameters.getParameter<double>("minZpos");
+      maxZpos = muonHistoParameters.getParameter<double>("maxZpos");
+      nintZpos = muonHistoParameters.getParameter<int>("nintZpos");
+      minPU = muonHistoParameters.getParameter<double>("minPU");
+      maxPU = muonHistoParameters.getParameter<double>("maxPU");
+      nintPU = muonHistoParameters.getParameter<int>("nintPU");
+      //
+      ptRes_rangeMin = muonHistoParameters.getParameter<double>("ptRes_rangeMin");
+      ptRes_rangeMax = muonHistoParameters.getParameter<double>("ptRes_rangeMax");
+      ptRes_nbin = muonHistoParameters.getParameter<int>("ptRes_nbin");
+      etaRes_rangeMin = muonHistoParameters.getParameter<double>("etaRes_rangeMin");
+      etaRes_rangeMax = muonHistoParameters.getParameter<double>("etaRes_rangeMax");
+      etaRes_nbin = muonHistoParameters.getParameter<int>("etaRes_nbin");
+      phiRes_rangeMin = muonHistoParameters.getParameter<double>("phiRes_rangeMin");
+      phiRes_rangeMax = muonHistoParameters.getParameter<double>("phiRes_rangeMax");
+      phiRes_nbin = muonHistoParameters.getParameter<int>("phiRes_nbin");
+      cotThetaRes_rangeMin = muonHistoParameters.getParameter<double>("cotThetaRes_rangeMin");
+      cotThetaRes_rangeMax = muonHistoParameters.getParameter<double>("cotThetaRes_rangeMax");
+      cotThetaRes_nbin = muonHistoParameters.getParameter<int>("cotThetaRes_nbin");
+      dxyRes_rangeMin = muonHistoParameters.getParameter<double>("dxyRes_rangeMin");
+      dxyRes_rangeMax = muonHistoParameters.getParameter<double>("dxyRes_rangeMax");
+      dxyRes_nbin = muonHistoParameters.getParameter<int>("dxyRes_nbin");
+      dzRes_rangeMin = muonHistoParameters.getParameter<double>("dzRes_rangeMin");
+      dzRes_rangeMax = muonHistoParameters.getParameter<double>("dzRes_rangeMax");
+      dzRes_nbin = muonHistoParameters.getParameter<int>("dzRes_nbin");
+      //
+      usetracker = muonHistoParameters.getParameter<bool>("usetracker");
+      usemuon = muonHistoParameters.getParameter<bool>("usemuon");
+      do_TRKhitsPlots = muonHistoParameters.getParameter<bool>("do_TRKhitsPlots");
+      do_MUOhitsPlots = muonHistoParameters.getParameter<bool>("do_MUOhitsPlots");
+      
+      if (useLogPt) {
+        minPt=log10(std::max(0.01,minPt));
+	maxPt=log10(maxPt);
       }
     }
   
   /// Destructor
   virtual ~MuonTrackValidatorBase() noexcept(false) { }
+ 
+  template<typename T> void fillPlotNoFlow (MonitorElement* h, T val) {
+    h->Fill(std::min(std::max(val,((T) h->getTH1()->GetXaxis()->GetXmin())),((T) h->getTH1()->GetXaxis()->GetXmax())));
+  }
   
-  virtual void doProfileX(TH2 * th2, MonitorElement* me){
+  void doProfileX(TH2 * th2, MonitorElement* me){
     if (th2->GetNbinsX()==me->getNbinsX()){
       TProfile * p1 = (TProfile*) th2->ProfileX();
       p1->Copy(*me->getTProfile());
@@ -123,47 +159,23 @@ class MuonTrackValidatorBase {
     }
   }
 
-  virtual void doProfileX(MonitorElement * th2m, MonitorElement* me) {
+  void doProfileX(MonitorElement * th2m, MonitorElement* me) {
     doProfileX(th2m->getTH2F(), me);
   }
 
-  virtual double getEta(double eta) {
-    if (useFabs) return fabs(eta);
+  //  virtual double getEta(double eta) {
+  double getEta(double eta) {
+    if (useFabsEta) return fabs(eta);
     else return eta;
   }
 
-  virtual double getPt(double pt) {
+  //  virtual double getPt(double pt) {
+  double getPt(double pt) {
     if (useInvPt && pt!=0) return 1/pt;
     else return pt;
   }
   
-  void fillPlotFromVector(MonitorElement* h, std::vector<int>& vec) {
-    for (unsigned int j=0; j<vec.size(); j++){
-      h->setBinContent(j+1, vec[j]);
-    }
-  }
-
-  void fillPlotFromVectors(MonitorElement* h, std::vector<int>& numerator, std::vector<int>& denominator,std::string type){
-    double value,err;
-    for (unsigned int j=0; j<numerator.size(); j++){
-      if (denominator[j]!=0){
-if (type=="effic")
-value = ((double) numerator[j])/((double) denominator[j]);
-else if (type=="fakerate")
-value = 1-((double) numerator[j])/((double) denominator[j]);
-else return;
-err = sqrt( value*(1-value)/(double) denominator[j] );
-h->setBinContent(j+1, value);
-h->setBinError(j+1,err);
-      }
-      else {
-h->setBinContent(j+1, 0);
-      }
-    }
-  }
-
-  void BinLogX(TH1*h)
-  {
+  void BinLogX(TH1*h) {
     
     TAxis *axis = h->GetXaxis();
     int bins = axis->GetNbins();
@@ -181,176 +193,13 @@ h->setBinContent(j+1, 0);
     delete[] new_bins;
   }
 
-  void setUpVectors() {
-    std::vector<double> etaintervalsv;
-    std::vector<double> phiintervalsv;
-    std::vector<double> pTintervalsv;
-    std::vector<double> dxyintervalsv;
-    std::vector<double> dzintervalsv;
-    std::vector<double> vertposintervalsv;
-    std::vector<double> zposintervalsv;
-    std::vector<int> totSIMveta,totASSveta,totASS2veta,totRECveta;
-    std::vector<int> totSIMvpT,totASSvpT,totASS2vpT,totRECvpT;
-    std::vector<int> totSIMv_hit,totASSv_hit,totASS2v_hit,totRECv_hit;
-    std::vector<int> totSIMv_phi,totASSv_phi,totASS2v_phi,totRECv_phi;
-    std::vector<int> totSIMv_dxy,totASSv_dxy,totASS2v_dxy,totRECv_dxy;
-    std::vector<int> totSIMv_dz,totASSv_dz,totASS2v_dz,totRECv_dz;
-    std::vector<int> totSIMv_vertpos,totASSv_vertpos,totSIMv_zpos,totASSv_zpos;
-
-    // for muon Validation
-    std::vector<int> totASSveta_Quality05, totASSveta_Quality075;
-    std::vector<int> totASSvpT_Quality05, totASSvpT_Quality075;
-    std::vector<int> totASSv_phi_Quality05, totASSv_phi_Quality075;
-
-    double step=(max-min)/nint;
-    std::ostringstream title,name;
-    etaintervalsv.push_back(min);
-    for (int k=1;k<nint+1;k++) {
-      double d=min+k*step;
-      etaintervalsv.push_back(d);
-      totSIMveta.push_back(0);
-      totASSveta.push_back(0);
-      totASS2veta.push_back(0);
-      totRECveta.push_back(0);
-      //
-      totASSveta_Quality05.push_back(0);
-      totASSveta_Quality075.push_back(0);
-    }
-    etaintervals.push_back(etaintervalsv);
-    totSIMeta.push_back(totSIMveta);
-    totASSeta.push_back(totASSveta);
-    totASS2eta.push_back(totASS2veta);
-    totRECeta.push_back(totRECveta);
-    //
-    totASSeta_Quality05.push_back(totASSveta_Quality05);
-    totASSeta_Quality075.push_back(totASSveta_Quality075);
-  
-    double steppT = (maxpT-minpT)/nintpT;
-    pTintervalsv.push_back(minpT);
-    for (int k=1;k<nintpT+1;k++) {
-      double d=0;
-      if(useLogPt)d=pow(10,minpT+k*steppT);
-      else d=minpT+k*steppT;
-      pTintervalsv.push_back(d);
-      totSIMvpT.push_back(0);
-      totASSvpT.push_back(0);
-      totASS2vpT.push_back(0);
-      totRECvpT.push_back(0);
-      //
-      totASSvpT_Quality05.push_back(0);
-      totASSvpT_Quality075.push_back(0);
-    }
-    pTintervals.push_back(pTintervalsv);
-    totSIMpT.push_back(totSIMvpT);
-    totASSpT.push_back(totASSvpT);
-    totASS2pT.push_back(totASS2vpT);
-    totRECpT.push_back(totRECvpT);
-    //
-    totASSpT_Quality05.push_back(totASSvpT_Quality05);
-    totASSpT_Quality075.push_back(totASSvpT_Quality075);
-
-    for (int k=1;k<nintHit+1;k++) {
-      totSIMv_hit.push_back(0);
-      totASSv_hit.push_back(0);
-      totASS2v_hit.push_back(0);
-      totRECv_hit.push_back(0);
-    }
-    totSIM_hit.push_back(totSIMv_hit);
-    totASS_hit.push_back(totASSv_hit);
-    totASS2_hit.push_back(totASS2v_hit);
-    totREC_hit.push_back(totRECv_hit);
-
-    double stepPhi = (maxPhi-minPhi)/nintPhi;
-    phiintervalsv.push_back(minPhi);
-    for (int k=1;k<nintPhi+1;k++) {
-      double d=minPhi+k*stepPhi;
-      phiintervalsv.push_back(d);
-      totSIMv_phi.push_back(0);
-      totASSv_phi.push_back(0);
-      totASS2v_phi.push_back(0);
-      totRECv_phi.push_back(0);
-      //
-      totASSv_phi_Quality05.push_back(0);
-      totASSv_phi_Quality075.push_back(0);
-    }
-    phiintervals.push_back(phiintervalsv);
-    totSIM_phi.push_back(totSIMv_phi);
-    totASS_phi.push_back(totASSv_phi);
-    totASS2_phi.push_back(totASS2v_phi);
-    totREC_phi.push_back(totRECv_phi);
-    //
-    totASS_phi_Quality05.push_back(totASSv_phi_Quality05);
-    totASS_phi_Quality075.push_back(totASSv_phi_Quality075);
-
-    double stepDxy = (maxDxy-minDxy)/nintDxy;
-    dxyintervalsv.push_back(minDxy);
-    for (int k=1;k<nintDxy+1;k++) {
-      double d=minDxy+k*stepDxy;
-      dxyintervalsv.push_back(d);
-      totSIMv_dxy.push_back(0);
-      totASSv_dxy.push_back(0);
-      totASS2v_dxy.push_back(0);
-      totRECv_dxy.push_back(0);
-    }
-    dxyintervals.push_back(dxyintervalsv);
-    totSIM_dxy.push_back(totSIMv_dxy);
-    totASS_dxy.push_back(totASSv_dxy);
-    totASS2_dxy.push_back(totASS2v_dxy);
-    totREC_dxy.push_back(totRECv_dxy);
-
-
-    double stepDz = (maxDz-minDz)/nintDz;
-    dzintervalsv.push_back(minDz);
-    for (int k=1;k<nintDz+1;k++) {
-      double d=minDz+k*stepDz;
-      dzintervalsv.push_back(d);
-      totSIMv_dz.push_back(0);
-      totASSv_dz.push_back(0);
-      totASS2v_dz.push_back(0);
-      totRECv_dz.push_back(0);
-    }
-    dzintervals.push_back(dzintervalsv);
-    totSIM_dz.push_back(totSIMv_dz);
-    totASS_dz.push_back(totASSv_dz);
-    totASS2_dz.push_back(totASS2v_dz);
-    totREC_dz.push_back(totRECv_dz);
-
-    double stepVertpos = (maxVertpos-minVertpos)/nintVertpos;
-    vertposintervalsv.push_back(minVertpos);
-    for (int k=1;k<nintVertpos+1;k++) {
-      double d=minVertpos+k*stepVertpos;
-      vertposintervalsv.push_back(d);
-      totSIMv_vertpos.push_back(0);
-      totASSv_vertpos.push_back(0);
-    }
-    vertposintervals.push_back(vertposintervalsv);
-    totSIM_vertpos.push_back(totSIMv_vertpos);
-    totASS_vertpos.push_back(totASSv_vertpos);
-      
-    double stepZpos = (maxZpos-minZpos)/nintZpos;
-    zposintervalsv.push_back(minZpos);
-    for (int k=1;k<nintZpos+1;k++) {
-      double d=minZpos+k*stepZpos;
-      zposintervalsv.push_back(d);
-      totSIMv_zpos.push_back(0);
-      totASSv_zpos.push_back(0);
-    }
-    zposintervals.push_back(zposintervalsv);
-    totSIM_zpos.push_back(totSIMv_zpos);
-    totASS_zpos.push_back(totASSv_zpos);
-
-  }
-
  protected:
 
-  DQMStore* dbe_;
-
   std::vector<edm::InputTag> label;
-  bool usetracker;
-  bool usemuon;
   edm::InputTag bsSrc;
   edm::InputTag label_tp_effic;
   edm::InputTag label_tp_fake;
+  edm::InputTag label_pileupinfo;
   std::vector<std::string> associators;
   std::string out;
   std::string parametersDefiner;
@@ -358,90 +207,73 @@ h->setBinContent(j+1, 0);
   edm::EDGetTokenT<reco::BeamSpot> bsSrc_Token;
   edm::EDGetTokenT<TrackingParticleCollection> tp_effic_Token;
   edm::EDGetTokenT<TrackingParticleCollection> tp_fake_Token;
-       
-  double min, max;
-  int nint;
-  bool useFabs;
-  double minpT, maxpT;
-  int nintpT;
-  double minHit, maxHit;
-  int nintHit;
-  double minPhi, maxPhi;
-  int nintPhi;
-  double minDxy, maxDxy;
-  int nintDxy;
-  double minDz, maxDz;
-  int nintDz;
-  double minVertpos, maxVertpos;
-  int nintVertpos;
-  double minZpos, maxZpos;
-  int nintZpos;
-  bool useInvPt;
-  //
-  double ptRes_rangeMin,ptRes_rangeMax,
-    phiRes_rangeMin,phiRes_rangeMax, cotThetaRes_rangeMin,cotThetaRes_rangeMax,
-    dxyRes_rangeMin,dxyRes_rangeMax, dzRes_rangeMin,dzRes_rangeMax;
-  int ptRes_nbin, cotThetaRes_nbin, phiRes_nbin, dxyRes_nbin, dzRes_nbin;
-  bool ignoremissingtkcollection_;
-  bool useLogPt;
-
+  edm::EDGetTokenT<std::vector<PileupSummaryInfo> > pileupinfo_Token;
   edm::ESHandle<MagneticField> theMF;
 
-  //sim
-  std::vector<MonitorElement*> h_ptSIM, h_etaSIM, h_tracksSIM, h_vertposSIM;
+  edm::ParameterSet muonHistoParameters;
+
+  double minEta, maxEta;  int nintEta;  bool useFabsEta;
+  double minPt, maxPt;  int nintPt;  bool useLogPt;  bool useInvPt; 
+  double minNHit, maxNHit;  int nintNHit;
+  double minDTHit, maxDTHit; int nintDTHit;
+  double minCSCHit, maxCSCHit;  int nintCSCHit;
+  double minRPCHit, maxRPCHit;  int nintRPCHit;
+  double minLayers, maxLayers;  int nintLayers;
+  double minPixels, maxPixels;  int nintPixels;
+  double minPhi, maxPhi;  int nintPhi;
+  double minDxy, maxDxy;  int nintDxy;
+  double minDz, maxDz;  int nintDz;
+  double minRpos, maxRpos;  int nintRpos;
+  double minZpos, maxZpos;  int nintZpos;
+  double minPU, maxPU;  int nintPU;
+  //
+  double ptRes_rangeMin,ptRes_rangeMax; int ptRes_nbin;
+  double etaRes_rangeMin,etaRes_rangeMax; int etaRes_nbin;
+  double phiRes_rangeMin,phiRes_rangeMax; int phiRes_nbin;
+  double cotThetaRes_rangeMin,cotThetaRes_rangeMax; int cotThetaRes_nbin;
+  double dxyRes_rangeMin,dxyRes_rangeMax; int dxyRes_nbin;
+  double dzRes_rangeMin,dzRes_rangeMax; int dzRes_nbin;
+      
+  bool usetracker, usemuon;
+  bool do_TRKhitsPlots, do_MUOhitsPlots;
+  bool ignoremissingtkcollection_;
 
   //1D
-  std::vector<MonitorElement*> h_tracks, h_fakes, h_hits, h_charge;
-  std::vector<MonitorElement*> h_recoeta, h_assoceta, h_assoc2eta, h_simuleta;
-  std::vector<MonitorElement*> h_recopT, h_assocpT, h_assoc2pT, h_simulpT;
-  std::vector<MonitorElement*> h_recohit, h_assochit, h_assoc2hit, h_simulhit;
-  std::vector<MonitorElement*> h_recophi, h_assocphi, h_assoc2phi, h_simulphi;
-  std::vector<MonitorElement*> h_recodxy, h_assocdxy, h_assoc2dxy, h_simuldxy;
-  std::vector<MonitorElement*> h_recodz, h_assocdz, h_assoc2dz, h_simuldz;
-  std::vector<MonitorElement*> h_assocvertpos, h_simulvertpos, h_assoczpos, h_simulzpos;
-  std::vector<MonitorElement*> h_pt, h_eta, h_pullTheta,h_pullPhi,h_pullDxy,h_pullDz,h_pullQoverp;
+  std::vector<MonitorElement*> h_tracks, h_fakes, h_nhits, h_charge;
+  std::vector<MonitorElement*> h_recoeta, h_assoceta, h_assoc2eta, h_simuleta, h_misideta;
+  std::vector<MonitorElement*> h_recopT, h_assocpT, h_assoc2pT, h_simulpT, h_misidpT;
+  std::vector<MonitorElement*> h_recohit, h_assochit, h_assoc2hit, h_simulhit, h_misidhit;
+  std::vector<MonitorElement*> h_recophi, h_assocphi, h_assoc2phi, h_simulphi, h_misidphi;
+  std::vector<MonitorElement*> h_recodxy, h_assocdxy, h_assoc2dxy, h_simuldxy, h_misiddxy;
+  std::vector<MonitorElement*> h_recodz, h_assocdz, h_assoc2dz, h_simuldz, h_misiddz;
+  std::vector<MonitorElement*> h_recopu, h_assocpu, h_assoc2pu, h_simulpu, h_misidpu;
 
-  std::vector<MonitorElement*> h_assoceta_Quality05, h_assoceta_Quality075;
-  std::vector<MonitorElement*> h_assocpT_Quality05, h_assocpT_Quality075;
-  std::vector<MonitorElement*> h_assocphi_Quality05, h_assocphi_Quality075;
+  std::vector<MonitorElement*> h_assocRpos, h_simulRpos, h_assocZpos, h_simulZpos;
+  std::vector<MonitorElement*> h_etaRes;
 
-
-  //2D
-  std::vector<MonitorElement*> nrec_vs_nsim;
-  std::vector<MonitorElement*> nrecHit_vs_nsimHit_sim2rec;
-  std::vector<MonitorElement*> nrecHit_vs_nsimHit_rec2sim;
-
-  //assoc hits
+  std::vector<MonitorElement*> h_nchi2, h_nchi2_prob, h_losthits;
+  std::vector<MonitorElement*> h_nmisslayers_inner,h_nmisslayers_outer,h_nlosthits;
+  std::vector<MonitorElement*> h_assochi2, h_assochi2_prob;
   std::vector<MonitorElement*> h_assocFraction, h_assocSharedHit;
-
-  //#hit vs eta: to be used with doProfileX
-  std::vector<MonitorElement*> nhits_vs_eta,
-    nDThits_vs_eta,nCSChits_vs_eta,nRPChits_vs_eta,nGEMhits_vs_eta,nME0hits_vs_eta;
-
-  std::vector<MonitorElement*> h_hits_eta,
-    h_DThits_eta,h_CSChits_eta,h_RPChits_eta,h_GEMhits_eta,h_ME0hits_eta;
-    
-
-  std::vector< std::vector<double> > etaintervals;
-  std::vector< std::vector<double> > pTintervals;
-  std::vector< std::vector<double> > phiintervals;
-  std::vector< std::vector<double> > dxyintervals;
-  std::vector< std::vector<double> > dzintervals;
-  std::vector< std::vector<double> > vertposintervals;
-  std::vector< std::vector<double> > zposintervals;
-  std::vector< std::vector<int> > totSIMeta,totRECeta,totASSeta,totASS2eta;
-  std::vector< std::vector<int> > totSIMpT,totRECpT,totASSpT,totASS2pT;
-  std::vector< std::vector<int> > totSIM_hit,totREC_hit,totASS_hit,totASS2_hit;
-  std::vector< std::vector<int> > totSIM_phi,totREC_phi,totASS_phi,totASS2_phi;
-  std::vector< std::vector<int> > totSIM_dxy,totREC_dxy,totASS_dxy,totASS2_dxy;
-  std::vector< std::vector<int> > totSIM_dz,totREC_dz,totASS_dz,totASS2_dz;
-  std::vector< std::vector<int> > totSIM_vertpos,totASS_vertpos,totSIM_zpos,totASS_zpos;
-
-  // for muon Validation (SimToReco distributions for Quality > 0.5, 0.75)
+  
+  //2D
+  std::vector<MonitorElement*> nRecHits_vs_nSimHits;
   std::vector<MonitorElement*> h_PurityVsQuality;
-  std::vector< std::vector<int> > totASSeta_Quality05,totASSeta_Quality075;
-  std::vector< std::vector<int> > totASSpT_Quality05, totASSpT_Quality075;
-  std::vector< std::vector<int> > totASS_phi_Quality05, totASS_phi_Quality075;
+  std::vector<MonitorElement*> chi2_vs_nhits, etares_vs_eta;
+  std::vector<MonitorElement*> ptres_vs_phi, chi2_vs_phi, nhits_vs_phi, phires_vs_phi;
+
+  std::vector<MonitorElement*> nhits_vs_eta,nDThits_vs_eta,nCSChits_vs_eta,nRPChits_vs_eta,nGEMhits_vs_eta,nME0hits_vs_eta;
+  std::vector<MonitorElement*> chi2_vs_eta, nlosthits_vs_eta;
+  std::vector<MonitorElement*> nTRK_LayersWithMeas_vs_eta,nPixel_LayersWithMeas_vs_eta;
+
+  std::vector<MonitorElement*> dxyres_vs_eta, ptres_vs_eta, dzres_vs_eta, phires_vs_eta, thetaCotres_vs_eta;
+  std::vector<MonitorElement*> dxyres_vs_pt, ptres_vs_pt, dzres_vs_pt, phires_vs_pt, thetaCotres_vs_pt;
+
+  std::vector<MonitorElement*> dxypull_vs_eta, ptpull_vs_eta, dzpull_vs_eta, phipull_vs_eta, thetapull_vs_eta;
+  std::vector<MonitorElement*> ptpull_vs_phi, phipull_vs_phi, thetapull_vs_phi;
+  std::vector<MonitorElement*> h_dxypulleta, h_ptpulleta, h_dzpulleta, h_phipulleta, h_thetapulleta;
+  std::vector<MonitorElement*> h_ptpullphi, h_phipullphi, h_thetapullphi;
+  std::vector<MonitorElement*> h_ptpull, h_qoverppull, h_thetapull, h_phipull, h_dxypull, h_dzpull;
 
 };
