+++ Validation/RecoMuon/plugins/SealModule.cc
@@ -2,10 +2,8 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 
 #include "Validation/RecoMuon/plugins/MuonTrackValidator.h"
-#include "Validation/RecoMuon/plugins/NewMuonTrackValidator.h"
 
 DEFINE_FWK_MODULE(MuonTrackValidator);
-DEFINE_FWK_MODULE(NewMuonTrackValidator);
