+++ Validation/RecoMuon/python/muonValidation_cff.py
@@ -1,417 +1,178 @@
+#
+# Production configuration for FullSim: muon track validation using MuonAssociatorByHits
+#
 import FWCore.ParameterSet.Config as cms
 
 from Validation.RecoMuon.selectors_cff import *
 from Validation.RecoMuon.associators_cff import *
-# Configurations for MuonTrackValidators
+from Validation.RecoMuon.histoParameters_cff import *
 import Validation.RecoMuon.MuonTrackValidator_cfi
+
 from SimTracker.TrackAssociation.LhcParametersDefinerForTP_cfi import *
 from SimTracker.TrackAssociation.CosmicParametersDefinerForTP_cfi import *
+
+from Validation.RecoMuon.RecoMuonValidator_cff import *
+
+# quickTrackAssociatorByHits on probeTracks used as monitor wrt MuonAssociatorByHits
 
 trkMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 trkMuonTrackVTrackAssoc.associatormap = 'tpToTkmuTrackAssociation'
 trkMuonTrackVTrackAssoc.associators = ('trackAssociatorByHits',)
 #trkMuonTrackVTrackAssoc.label = ('generalTracks',)
 trkMuonTrackVTrackAssoc.label = ('probeTracks',)
-trkMuonTrackVTrackAssoc.usetracker = True
-trkMuonTrackVTrackAssoc.usemuon = False
+trkMuonTrackVTrackAssoc.muonHistoParameters = trkMuonHistoParameters
 
-trkCosmicMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-trkCosmicMuonTrackVTrackAssoc.associatormap = 'tpToTkCosmicTrackAssociation'
-trkCosmicMuonTrackVTrackAssoc.associators = ('trackAssociatorByHits',)
-trkCosmicMuonTrackVTrackAssoc.label = ('ctfWithMaterialTracksP5LHCNavigation',)
-trkCosmicMuonTrackVTrackAssoc.usetracker = True
-trkCosmicMuonTrackVTrackAssoc.usemuon = False
-
-staMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-staMuonTrackVTrackAssoc.associatormap = 'tpToStaTrackAssociation'
-staMuonTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-staMuonTrackVTrackAssoc.label = ('standAloneMuons',)
-staMuonTrackVTrackAssoc.usetracker = False
-staMuonTrackVTrackAssoc.usemuon = True
-
-staUpdMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-staUpdMuonTrackVTrackAssoc.associatormap = 'tpToStaUpdTrackAssociation'
-staUpdMuonTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-staUpdMuonTrackVTrackAssoc.label = ('standAloneMuons:UpdatedAtVtx',)
-staUpdMuonTrackVTrackAssoc.usetracker = False
-staUpdMuonTrackVTrackAssoc.usemuon = True
-
-glbMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-glbMuonTrackVTrackAssoc.associatormap = 'tpToGlbTrackAssociation'
-glbMuonTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-glbMuonTrackVTrackAssoc.label = ('globalMuons',)
-glbMuonTrackVTrackAssoc.usetracker = True
-glbMuonTrackVTrackAssoc.usemuon = True
-
-tevMuonFirstTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-tevMuonFirstTrackVTrackAssoc.associatormap = 'tpToTevFirstTrackAssociation'
-tevMuonFirstTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-tevMuonFirstTrackVTrackAssoc.label = ('tevMuons:firstHit',)
-tevMuonFirstTrackVTrackAssoc.usetracker = True
-tevMuonFirstTrackVTrackAssoc.usemuon = True
-
-tevMuonPickyTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-tevMuonPickyTrackVTrackAssoc.associatormap = 'tpToTevPickyTrackAssociation'
-tevMuonPickyTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-tevMuonPickyTrackVTrackAssoc.label = ('tevMuons:picky',)
-tevMuonPickyTrackVTrackAssoc.usetracker = True
-tevMuonPickyTrackVTrackAssoc.usemuon = True
-
-tevMuonDytTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-tevMuonDytTrackVTrackAssoc.associatormap = 'tpToTevDytTrackAssociation'
-tevMuonDytTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-tevMuonDytTrackVTrackAssoc.label = ('tevMuons:dyt',)
-tevMuonDytTrackVTrackAssoc.usetracker = True
-tevMuonDytTrackVTrackAssoc.usemuon = True
-
-staCosmicMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-staCosmicMuonTrackVTrackAssoc.associatormap = 'tpToStaCosmicTrackAssociation'
-staCosmicMuonTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-staCosmicMuonTrackVTrackAssoc.label = ('cosmicMuons',)
-staCosmicMuonTrackVTrackAssoc.usetracker = False
-staCosmicMuonTrackVTrackAssoc.usemuon = True
-
-glbCosmicMuonTrackVTrackAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-glbCosmicMuonTrackVTrackAssoc.associatormap = 'tpToGlbCosmicTrackAssociation'
-glbCosmicMuonTrackVTrackAssoc.associators = ('trackAssociatorByDeltaR',)
-glbCosmicMuonTrackVTrackAssoc.label = ('globalCosmicMuons',)
-glbCosmicMuonTrackVTrackAssoc.usetracker = True
-glbCosmicMuonTrackVTrackAssoc.usemuon = True
+# MuonAssociatorByHits used for all track collections
 
 trkProbeTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-#trkMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
-trkProbeTrackVMuonAssoc.associatormap = 'tpToTkMuonAssociation' 
-trkProbeTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
-##trkMuonTrackVMuonAssoc.label = ('generalTracks',)
+trkProbeTrackVMuonAssoc.associatormap = 'tpToTkMuonAssociation'
+#trkProbeTrackVMuonAssoc.label = ('generalTracks',)
 trkProbeTrackVMuonAssoc.label = ('probeTracks',)
-trkProbeTrackVMuonAssoc.usetracker = True
-trkProbeTrackVMuonAssoc.usemuon = False
+trkProbeTrackVMuonAssoc.muonHistoParameters = trkMuonHistoParameters
 
 staSeedTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staSeedTrackVMuonAssoc.associatormap = 'tpToStaSeedAssociation'
-staSeedTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 staSeedTrackVMuonAssoc.label = ('seedsOfSTAmuons',)
-staSeedTrackVMuonAssoc.usetracker = False
-staSeedTrackVMuonAssoc.usemuon = True
+staSeedTrackVMuonAssoc.muonHistoParameters = staSeedMuonHistoParameters
 
 staMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staMuonTrackVMuonAssoc.associatormap = 'tpToStaMuonAssociation'
-staMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 staMuonTrackVMuonAssoc.label = ('standAloneMuons',)
-staMuonTrackVMuonAssoc.usetracker = False
-staMuonTrackVMuonAssoc.usemuon = True
+staMuonTrackVMuonAssoc.muonHistoParameters = staMuonHistoParameters
 
 staUpdMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staUpdMuonTrackVMuonAssoc.associatormap = 'tpToStaUpdMuonAssociation'
-staUpdMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 staUpdMuonTrackVMuonAssoc.label = ('standAloneMuons:UpdatedAtVtx',)
-staUpdMuonTrackVMuonAssoc.usetracker = False
-staUpdMuonTrackVMuonAssoc.usemuon = True
+staUpdMuonTrackVMuonAssoc.muonHistoParameters = staUpdMuonHistoParameters
 
 glbMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 glbMuonTrackVMuonAssoc.associatormap = 'tpToGlbMuonAssociation'
-glbMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
-glbMuonTrackVMuonAssoc.label = ('extractedGlobalMuons',)
-glbMuonTrackVMuonAssoc.usetracker = True
-glbMuonTrackVMuonAssoc.usemuon = True
+glbMuonTrackVMuonAssoc.label = ('globalMuons',)
+glbMuonTrackVMuonAssoc.muonHistoParameters = glbMuonHistoParameters
 
 staRefitMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staRefitMuonTrackVMuonAssoc.associatormap = 'tpToStaRefitMuonAssociation'
-staRefitMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 staRefitMuonTrackVMuonAssoc.label = ('refittedStandAloneMuons',)
-staRefitMuonTrackVMuonAssoc.usetracker = False
-staRefitMuonTrackVMuonAssoc.usemuon = True
+staRefitMuonTrackVMuonAssoc.muonHistoParameters = staMuonHistoParameters
 
 staRefitUpdMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staRefitUpdMuonTrackVMuonAssoc.associatormap = 'tpToStaRefitUpdMuonAssociation'
-staRefitUpdMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 staRefitUpdMuonTrackVMuonAssoc.label = ('refittedStandAloneMuons:UpdatedAtVtx',)
-staRefitUpdMuonTrackVMuonAssoc.usetracker = False
-staRefitUpdMuonTrackVMuonAssoc.usemuon = True
+staRefitUpdMuonTrackVMuonAssoc.muonHistoParameters = staUpdMuonHistoParameters
 
 displacedTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 displacedTrackVMuonAssoc.associatormap = 'tpToDisplacedTrkMuonAssociation'
-displacedTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 displacedTrackVMuonAssoc.label = ('displacedTracks',)
-displacedTrackVMuonAssoc.usetracker = True
-displacedTrackVMuonAssoc.usemuon = False
-displacedTrackVMuonAssoc.tipTP = cms.double(85.)
-displacedTrackVMuonAssoc.lipTP = cms.double(210.)
+displacedTrackVMuonAssoc.muonTPSelector = displacedMuonTPSet
+displacedTrackVMuonAssoc.muonHistoParameters = displacedTrkMuonHistoParameters
 
 displacedStaSeedTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 displacedStaSeedTrackVMuonAssoc.associatormap = 'tpToDisplacedStaSeedAssociation'
-displacedStaSeedTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 displacedStaSeedTrackVMuonAssoc.label = ('seedsOfDisplacedSTAmuons',)
-displacedStaSeedTrackVMuonAssoc.usetracker = False
-displacedStaSeedTrackVMuonAssoc.usemuon = True
-displacedStaSeedTrackVMuonAssoc.tipTP = cms.double(85.)
-displacedStaSeedTrackVMuonAssoc.lipTP = cms.double(210.)
+displacedStaSeedTrackVMuonAssoc.muonTPSelector = displacedMuonTPSet
+displacedStaSeedTrackVMuonAssoc.muonHistoParameters = displacedStaSeedMuonHistoParameters
 
 displacedStaMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 displacedStaMuonTrackVMuonAssoc.associatormap = 'tpToDisplacedStaMuonAssociation'
-displacedStaMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 displacedStaMuonTrackVMuonAssoc.label = ('displacedStandAloneMuons',)
-displacedStaMuonTrackVMuonAssoc.usetracker = False
-displacedStaMuonTrackVMuonAssoc.usemuon = True
-displacedStaMuonTrackVMuonAssoc.tipTP = cms.double(85.)
-displacedStaMuonTrackVMuonAssoc.lipTP = cms.double(210.)
+displacedStaMuonTrackVMuonAssoc.muonTPSelector = displacedMuonTPSet
+displacedStaMuonTrackVMuonAssoc.muonHistoParameters = displacedStaMuonHistoParameters
 
 displacedGlbMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 displacedGlbMuonTrackVMuonAssoc.associatormap = 'tpToDisplacedGlbMuonAssociation'
-displacedGlbMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 displacedGlbMuonTrackVMuonAssoc.label = ('displacedGlobalMuons',)
-displacedGlbMuonTrackVMuonAssoc.usetracker = True
-displacedGlbMuonTrackVMuonAssoc.usemuon = True
-displacedGlbMuonTrackVMuonAssoc.tipTP = cms.double(85.)
-displacedGlbMuonTrackVMuonAssoc.lipTP = cms.double(210.)
+displacedGlbMuonTrackVMuonAssoc.muonTPSelector = displacedMuonTPSet
+displacedGlbMuonTrackVMuonAssoc.muonHistoParameters = displacedGlbMuonHistoParameters
 
 tevMuonFirstTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 tevMuonFirstTrackVMuonAssoc.associatormap = 'tpToTevFirstMuonAssociation'
-tevMuonFirstTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 tevMuonFirstTrackVMuonAssoc.label = ('tevMuons:firstHit',)
-tevMuonFirstTrackVMuonAssoc.usetracker = True
-tevMuonFirstTrackVMuonAssoc.usemuon = True
+tevMuonFirstTrackVMuonAssoc.muonHistoParameters = glbMuonHistoParameters
 
 tevMuonPickyTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 tevMuonPickyTrackVMuonAssoc.associatormap = 'tpToTevPickyMuonAssociation'
-tevMuonPickyTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 tevMuonPickyTrackVMuonAssoc.label = ('tevMuons:picky',)
-tevMuonPickyTrackVMuonAssoc.usetracker = True
-tevMuonPickyTrackVMuonAssoc.usemuon = True
+tevMuonPickyTrackVMuonAssoc.muonHistoParameters = glbMuonHistoParameters
 
 tevMuonDytTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 tevMuonDytTrackVMuonAssoc.associatormap = 'tpToTevDytMuonAssociation'
-tevMuonDytTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 tevMuonDytTrackVMuonAssoc.label = ('tevMuons:dyt',)
-tevMuonDytTrackVMuonAssoc.usetracker = True
-tevMuonDytTrackVMuonAssoc.usemuon = True
+tevMuonDytTrackVMuonAssoc.muonHistoParameters = glbMuonHistoParameters
 
 gemMuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 gemMuonTrackVMuonAssoc.associatormap = 'tpToGEMMuonMuonAssociation'
-gemMuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 gemMuonTrackVMuonAssoc.label = ('extractGemMuons',)
-gemMuonTrackVMuonAssoc.minRapidityTP = -2.4
-gemMuonTrackVMuonAssoc.maxRapidityTP = +2.4
-gemMuonTrackVMuonAssoc.usetracker = True
-gemMuonTrackVMuonAssoc.usemuon = False
-gemMuonTrackVMuonAssoc.min = -2.4
-gemMuonTrackVMuonAssoc.max = +2.4
-gemMuonTrackVMuonAssoc.nint = 48
+gemMuonTrackVMuonAssoc.muonHistoParameters = gemMuonHistoParameters
 
 me0MuonTrackVMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 me0MuonTrackVMuonAssoc.associatormap = 'tpToME0MuonMuonAssociation'
-me0MuonTrackVMuonAssoc.associators = ('MuonAssociationByHits',)
 me0MuonTrackVMuonAssoc.label = ('extractMe0Muons',)
-me0MuonTrackVMuonAssoc.minRapidityTP = -2.8
-me0MuonTrackVMuonAssoc.maxRapidityTP = +2.8
-me0MuonTrackVMuonAssoc.usetracker = True
-me0MuonTrackVMuonAssoc.usemuon = False
-me0MuonTrackVMuonAssoc.min = -2.8
-me0MuonTrackVMuonAssoc.max = +2.8
-me0MuonTrackVMuonAssoc.nint = 56
+me0MuonTrackVMuonAssoc.muonTPSelector = me0MuonTPSet
+me0MuonTrackVMuonAssoc.muonHistoParameters = me0MuonHistoParameters
 
 # cosmics 2-leg reco
 trkCosmicMuonTrackVSelMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 trkCosmicMuonTrackVSelMuonAssoc.associatormap = 'tpToTkCosmicSelMuonAssociation'
-trkCosmicMuonTrackVSelMuonAssoc.associators = ('MuonAssociationByHits',)
 trkCosmicMuonTrackVSelMuonAssoc.label = ('ctfWithMaterialTracksP5LHCNavigation',)
-trkCosmicMuonTrackVSelMuonAssoc.usetracker = True
-trkCosmicMuonTrackVSelMuonAssoc.usemuon = False
 trkCosmicMuonTrackVSelMuonAssoc.parametersDefiner = cms.string('CosmicParametersDefinerForTP')
-trkCosmicMuonTrackVSelMuonAssoc.ptMinTP = cms.double(1.)
-trkCosmicMuonTrackVSelMuonAssoc.tipTP = cms.double(80.)
-trkCosmicMuonTrackVSelMuonAssoc.lipTP = cms.double(212.)
+trkCosmicMuonTrackVSelMuonAssoc.muonTPSelector = cosmicMuonTPSet
 trkCosmicMuonTrackVSelMuonAssoc.BiDirectional_RecoToSim_association = False
+trkCosmicMuonTrackVSelMuonAssoc.muonHistoParameters = trkCosmicMuonHistoParameters
 
 staCosmicMuonTrackVSelMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staCosmicMuonTrackVSelMuonAssoc.associatormap = 'tpToStaCosmicSelMuonAssociation'
-staCosmicMuonTrackVSelMuonAssoc.associators = ('MuonAssociationByHits',)
 staCosmicMuonTrackVSelMuonAssoc.label = ('cosmicMuons',)
-staCosmicMuonTrackVSelMuonAssoc.usetracker = False
-staCosmicMuonTrackVSelMuonAssoc.usemuon = True
 staCosmicMuonTrackVSelMuonAssoc.parametersDefiner = cms.string('CosmicParametersDefinerForTP')
-staCosmicMuonTrackVSelMuonAssoc.ptMinTP = cms.double(1.)
-staCosmicMuonTrackVSelMuonAssoc.tipTP = cms.double(80.)
-staCosmicMuonTrackVSelMuonAssoc.lipTP = cms.double(212.)
+staCosmicMuonTrackVSelMuonAssoc.muonTPSelector = cosmicMuonTPSet
 staCosmicMuonTrackVSelMuonAssoc.BiDirectional_RecoToSim_association = False
+staCosmicMuonTrackVSelMuonAssoc.muonHistoParameters = staCosmicMuonHistoParameters
 
 glbCosmicMuonTrackVSelMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 glbCosmicMuonTrackVSelMuonAssoc.associatormap = 'tpToGlbCosmicSelMuonAssociation'
-glbCosmicMuonTrackVSelMuonAssoc.associators = ('MuonAssociationByHits',)
 glbCosmicMuonTrackVSelMuonAssoc.label = ('globalCosmicMuons',)
-glbCosmicMuonTrackVSelMuonAssoc.usetracker = True
-glbCosmicMuonTrackVSelMuonAssoc.usemuon = True
 glbCosmicMuonTrackVSelMuonAssoc.parametersDefiner = cms.string('CosmicParametersDefinerForTP')
-glbCosmicMuonTrackVSelMuonAssoc.ptMinTP = cms.double(1.)
-glbCosmicMuonTrackVSelMuonAssoc.tipTP = cms.double(80.)
-glbCosmicMuonTrackVSelMuonAssoc.lipTP = cms.double(212.)
+glbCosmicMuonTrackVSelMuonAssoc.muonTPSelector = cosmicMuonTPSet
 glbCosmicMuonTrackVSelMuonAssoc.BiDirectional_RecoToSim_association = False
+glbCosmicMuonTrackVSelMuonAssoc.muonHistoParameters = glbCosmicMuonHistoParameters
 
 # cosmics 1-leg reco
 trkCosmic1LegMuonTrackVSelMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 trkCosmic1LegMuonTrackVSelMuonAssoc.associatormap = 'tpToTkCosmic1LegSelMuonAssociation'
-trkCosmic1LegMuonTrackVSelMuonAssoc.associators = ('MuonAssociationByHits',)
 trkCosmic1LegMuonTrackVSelMuonAssoc.label = ('ctfWithMaterialTracksP5',)
-trkCosmic1LegMuonTrackVSelMuonAssoc.usetracker = True
-trkCosmic1LegMuonTrackVSelMuonAssoc.usemuon = False
 trkCosmic1LegMuonTrackVSelMuonAssoc.parametersDefiner = cms.string('CosmicParametersDefinerForTP')
-trkCosmic1LegMuonTrackVSelMuonAssoc.ptMinTP = cms.double(1.)
-trkCosmic1LegMuonTrackVSelMuonAssoc.tipTP = cms.double(80.)
-trkCosmic1LegMuonTrackVSelMuonAssoc.lipTP = cms.double(212.)
-trkCosmic1LegMuonTrackVSelMuonAssoc.BiDirectional_RecoToSim_association = False
+trkCosmic1LegMuonTrackVSelMuonAssoc.muonTPSelector = cosmicMuonTPSet
+trkCosmic1LegMuonTrackVSelMuonAssoc.muonHistoParameters = trkCosmic1LegMuonHistoParameters
 
 staCosmic1LegMuonTrackVSelMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 staCosmic1LegMuonTrackVSelMuonAssoc.associatormap = 'tpToStaCosmic1LegSelMuonAssociation'
-staCosmic1LegMuonTrackVSelMuonAssoc.associators = ('MuonAssociationByHits',)
 staCosmic1LegMuonTrackVSelMuonAssoc.label = ('cosmicMuons1Leg',)
-staCosmic1LegMuonTrackVSelMuonAssoc.usetracker = False
-staCosmic1LegMuonTrackVSelMuonAssoc.usemuon = True
 staCosmic1LegMuonTrackVSelMuonAssoc.parametersDefiner = cms.string('CosmicParametersDefinerForTP')
-staCosmic1LegMuonTrackVSelMuonAssoc.ptMinTP = cms.double(1.)
-staCosmic1LegMuonTrackVSelMuonAssoc.tipTP = cms.double(80.)
-staCosmic1LegMuonTrackVSelMuonAssoc.lipTP = cms.double(212.)
-staCosmic1LegMuonTrackVSelMuonAssoc.BiDirectional_RecoToSim_association = False
+staCosmic1LegMuonTrackVSelMuonAssoc.muonTPSelector = cosmicMuonTPSet
+staCosmic1LegMuonTrackVSelMuonAssoc.muonHistoParameters = staCosmic1LegMuonHistoParameters
 
 glbCosmic1LegMuonTrackVSelMuonAssoc = Validation.RecoMuon.MuonTrackValidator_cfi.muonTrackValidator.clone()
 glbCosmic1LegMuonTrackVSelMuonAssoc.associatormap = 'tpToGlbCosmic1LegSelMuonAssociation'
-glbCosmic1LegMuonTrackVSelMuonAssoc.associators = ('MuonAssociationByHits',)
 glbCosmic1LegMuonTrackVSelMuonAssoc.label = ('globalCosmicMuons1Leg',)
-glbCosmic1LegMuonTrackVSelMuonAssoc.usetracker = True
-glbCosmic1LegMuonTrackVSelMuonAssoc.usemuon = True
 glbCosmic1LegMuonTrackVSelMuonAssoc.parametersDefiner = cms.string('CosmicParametersDefinerForTP')
-glbCosmic1LegMuonTrackVSelMuonAssoc.ptMinTP = cms.double(1.)
-glbCosmic1LegMuonTrackVSelMuonAssoc.tipTP = cms.double(80.)
-glbCosmic1LegMuonTrackVSelMuonAssoc.lipTP = cms.double(212.)
-glbCosmic1LegMuonTrackVSelMuonAssoc.BiDirectional_RecoToSim_association = False
+glbCosmic1LegMuonTrackVSelMuonAssoc.muonTPSelector = cosmicMuonTPSet
+glbCosmic1LegMuonTrackVSelMuonAssoc.muonHistoParameters = glbCosmic1LegMuonHistoParameters
 
-# Configurations for RecoMuonValidators
-from RecoMuon.TrackingTools.MuonServiceProxy_cff import *
-from Validation.RecoMuon.RecoMuonValidator_cfi import *
-
-#import SimGeneral.MixingModule.mixNoPU_cfi
-from SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi import *
-from SimMuon.MCTruth.MuonAssociatorByHits_cfi import muonAssociatorByHitsCommonParameters
-
-#tracker
-muonAssociatorByHitsNoSimHitsHelperTrk = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperTrk.UseTracker = True
-muonAssociatorByHitsNoSimHitsHelperTrk.UseMuon  = False
-recoMuonVMuAssoc_trk = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_trk.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_Trk'
-recoMuonVMuAssoc_trk.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_trk.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperTrk'
-recoMuonVMuAssoc_trk.trackType = 'inner'
-recoMuonVMuAssoc_trk.selection = "isTrackerMuon"
-
-#tracker and PF
-muonAssociatorByHitsNoSimHitsHelperTrkPF = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperTrkPF.UseTracker = True
-muonAssociatorByHitsNoSimHitsHelperTrkPF.UseMuon  = False
-recoMuonVMuAssoc_trkPF = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_trkPF.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_TrkPF'
-recoMuonVMuAssoc_trkPF.usePFMuon = True
-recoMuonVMuAssoc_trkPF.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_trkPF.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperTrkPF'
-recoMuonVMuAssoc_trkPF.trackType = 'inner'
-recoMuonVMuAssoc_trkPF.selection = "isTrackerMuon & isPFMuon"
-
-#standalone
-muonAssociatorByHitsNoSimHitsHelperStandalone = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperStandalone.UseTracker = False
-muonAssociatorByHitsNoSimHitsHelperStandalone.UseMuon  = True
-recoMuonVMuAssoc_sta = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_sta.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_Sta'
-recoMuonVMuAssoc_sta.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_sta.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperStandalone'
-recoMuonVMuAssoc_sta.trackType = 'outer'
-recoMuonVMuAssoc_sta.selection = "isStandAloneMuon"
-
-#seed of StandAlone
-muonAssociatorByHitsNoSimHitsHelperSeedStandalone = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperSeedStandalone.UseTracker = False
-muonAssociatorByHitsNoSimHitsHelperSeedStandalone.UseMuon  = True
-recoMuonVMuAssoc_seedSta = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_seedSta.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_SeedSta'
-recoMuonVMuAssoc_seedSta.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_seedSta.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperStandalone'
-recoMuonVMuAssoc_seedSta.trackType = 'outer'
-recoMuonVMuAssoc_seedSta.selection = ""
-
-#standalone and PF
-muonAssociatorByHitsNoSimHitsHelperStandalonePF = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperStandalonePF.UseTracker = False
-muonAssociatorByHitsNoSimHitsHelperStandalonePF.UseMuon  = True
-recoMuonVMuAssoc_staPF = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_staPF.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_StaPF'
-recoMuonVMuAssoc_staPF.usePFMuon = True
-recoMuonVMuAssoc_staPF.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_staPF.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperStandalonePF'
-recoMuonVMuAssoc_staPF.trackType = 'outer'
-recoMuonVMuAssoc_staPF.selection = "isStandAloneMuon & isPFMuon"
-
-#global
-muonAssociatorByHitsNoSimHitsHelperGlobal = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperGlobal.UseTracker = True
-muonAssociatorByHitsNoSimHitsHelperGlobal.UseMuon  = True
-recoMuonVMuAssoc_glb = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_glb.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_Glb'
-recoMuonVMuAssoc_glb.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_glb.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperGlobal'
-recoMuonVMuAssoc_glb.trackType = 'global'
-recoMuonVMuAssoc_glb.selection = "isGlobalMuon"
-
-#global and PF
-muonAssociatorByHitsNoSimHitsHelperGlobalPF = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperGlobalPF.UseTracker = True
-muonAssociatorByHitsNoSimHitsHelperGlobalPF.UseMuon  = True
-recoMuonVMuAssoc_glbPF = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_glbPF.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_GlbPF'
-recoMuonVMuAssoc_glbPF.usePFMuon = True
-recoMuonVMuAssoc_glbPF.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_glbPF.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperGlobalPF'
-recoMuonVMuAssoc_glbPF.trackType = 'global'
-recoMuonVMuAssoc_glbPF.selection = "isGlobalMuon & isPFMuon"
-
-#tight
-muonAssociatorByHitsNoSimHitsHelperTight = SimMuon.MCTruth.muonAssociatorByHitsNoSimHitsHelper_cfi.muonAssociatorByHitsNoSimHitsHelper.clone()
-muonAssociatorByHitsNoSimHitsHelperTight.UseTracker = True
-muonAssociatorByHitsNoSimHitsHelperTight.UseMuon  = True
-recoMuonVMuAssoc_tgt = Validation.RecoMuon.RecoMuonValidator_cfi.recoMuonValidator.clone()
-recoMuonVMuAssoc_tgt.subDir = 'Muons/RecoMuonV/RecoMuon_MuonAssoc_Tgt'
-recoMuonVMuAssoc_tgt.simLabel = 'mix:MergedTrackTruth'
-recoMuonVMuAssoc_tgt.muAssocLabel = 'muonAssociatorByHitsNoSimHitsHelperTight'
-recoMuonVMuAssoc_tgt.trackType = 'global'
-recoMuonVMuAssoc_tgt.selection = 'isGlobalMuon'
-recoMuonVMuAssoc_tgt.wantTightMuon = True
-recoMuonVMuAssoc_tgt.beamSpot = 'offlineBeamSpot'
-recoMuonVMuAssoc_tgt.primaryVertex = 'offlinePrimaryVertices'
-
-# Muon validation sequences
+##################################################################################
+# Muon validation sequences using MuonTrackValidator
+#
 muonValidation_seq = cms.Sequence(
     probeTracks_seq + tpToTkMuonAssociation + trkProbeTrackVMuonAssoc
     +trackAssociatorByHits + tpToTkmuTrackAssociation + trkMuonTrackVTrackAssoc
     +seedsOfSTAmuons_seq + tpToStaSeedAssociation + staSeedTrackVMuonAssoc
     +tpToStaMuonAssociation + staMuonTrackVMuonAssoc
     +tpToStaUpdMuonAssociation + staUpdMuonTrackVMuonAssoc
-    +extractedMuonTracks_seq + tpToGlbMuonAssociation + glbMuonTrackVMuonAssoc
-    +muonAssociatorByHitsNoSimHitsHelperTrk +recoMuonVMuAssoc_trk
-    +muonAssociatorByHitsNoSimHitsHelperStandalone +recoMuonVMuAssoc_sta
-    +muonAssociatorByHitsNoSimHitsHelperGlobal +recoMuonVMuAssoc_glb
-    +muonAssociatorByHitsNoSimHitsHelperTight +recoMuonVMuAssoc_tgt
+    +tpToGlbMuonAssociation + glbMuonTrackVMuonAssoc
 )
 
 muonValidation_reduced_seq = cms.Sequence(
     probeTracks_seq + tpToTkMuonAssociation + trkProbeTrackVMuonAssoc
-#    +trackAssociatorByHits + tpToTkmuTrackAssociation + trkMuonTrackVTrackAssoc
-#    +seedsOfSTAmuons_seq + tpToStaSeedAssociation + staSeedTrackVMuonAssoc
-#    +tpToStaMuonAssociation + staMuonTrackVMuonAssoc
     +tpToStaUpdMuonAssociation + staUpdMuonTrackVMuonAssoc
-    +extractedMuonTracks_seq + tpToGlbMuonAssociation + glbMuonTrackVMuonAssoc
-#    +muonAssociatorByHitsNoSimHitsHelperTrk +recoMuonVMuAssoc_trk
-#    +muonAssociatorByHitsNoSimHitsHelperStandalone +recoMuonVMuAssoc_sta
-#    +muonAssociatorByHitsNoSimHitsHelperGlobal +recoMuonVMuAssoc_glb
-#    +muonAssociatorByHitsNoSimHitsHelperTight +recoMuonVMuAssoc_tgt
-#    +seedsOfDisplacedSTAmuons_seq + tpToDisplacedStaSeedAssociation + displacedStaSeedTrackVMuonAssoc
+    +tpToGlbMuonAssociation + glbMuonTrackVMuonAssoc
     +tpToDisplacedStaMuonAssociation + displacedStaMuonTrackVMuonAssoc
     +tpToDisplacedTrkMuonAssociation + displacedTrackVMuonAssoc
     +tpToDisplacedGlbMuonAssociation + displacedGlbMuonTrackVMuonAssoc
@@ -444,34 +205,37 @@ muonValidationCosmic_seq = cms.Sequence(
     +tpToGlbCosmic1LegSelMuonAssociation + glbCosmic1LegMuonTrackVSelMuonAssoc
 )
 
-# The full offline muon validation sequence
-recoMuonValidation = cms.Sequence(
-    muonValidation_seq + muonValidationTEV_seq + muonValidationRefit_seq + muonValidationDisplaced_seq
-)
+gemMuonValidation = cms.Sequence(extractGemMuonsTracks_seq + tpToGEMMuonMuonAssociation + gemMuonTrackVMuonAssoc)
+me0MuonValidation = cms.Sequence(extractMe0MuonsTracks_seq + tpToME0MuonMuonAssociation + me0MuonTrackVMuonAssoc)
 
-# no displaces in fastsim
+##########################################################################
+# The full offline muon validation sequence
+#
+recoMuonValidation = cms.Sequence(
+    muonValidation_seq + muonValidationTEV_seq + muonValidationRefit_seq + muonValidationDisplaced_seq + muonValidationRMV_seq
+    )
+
+# no displaced muons in fastsim
 from Configuration.Eras.Modifier_fastSim_cff import fastSim
-fastSim.toReplaceWith(recoMuonValidation, cms.Sequence(muonValidation_seq + muonValidationTEV_seq + muonValidationRefit_seq))
+fastSim.toReplaceWith(recoMuonValidation, cms.Sequence(muonValidation_seq + muonValidationTEV_seq + muonValidationRefit_seq + muonValidationRMV_seq))
 
 # sequence for cosmic muons
 recoCosmicMuonValidation = cms.Sequence(
     muonValidationCosmic_seq
-)
+    )
 
-gemMuonValidation = cms.Sequence(extractGemMuonsTracks_seq + tpToGEMMuonMuonAssociation + gemMuonTrackVMuonAssoc)
-me0MuonValidation = cms.Sequence(extractMe0MuonsTracks_seq + tpToME0MuonMuonAssociation + me0MuonTrackVMuonAssoc)
-
+# sequences for muon upgrades
+#
 _run3_muonValidation = muonValidation_seq.copy() #For full validation
-#_run3_muonValidation = muonValidation_reduced_seq.copy()
+_run3_muonValidation = muonValidation_reduced_seq.copy()
 _run3_muonValidation += gemMuonValidation
 
-#_phase2_muonValidation = _run3_muonValidation.copy()
-_phase2_muonValidation = muonValidation_reduced_seq.copy()
+_phase2_muonValidation = _run3_muonValidation.copy()
 _phase2_muonValidation += me0MuonValidation
 
 from Configuration.Eras.Modifier_run3_GEM_cff import run3_GEM
 run3_GEM.toReplaceWith( muonValidation_seq, _run3_muonValidation ) #For full validation
-#run3_GEM.toReplaceWith( recoMuonValidation, _run3_muonValidation )
+run3_GEM.toReplaceWith( recoMuonValidation, _run3_muonValidation )
 from Configuration.Eras.Modifier_phase2_muon_cff import phase2_muon
-#phase2_muon.toReplaceWith( muonValidation_seq, _phase2_muonValidation ) #For full validation
+phase2_muon.toReplaceWith( muonValidation_seq, _phase2_muonValidation ) #For full validation
 phase2_muon.toReplaceWith( recoMuonValidation, _phase2_muonValidation )
