+++ Validation/RecoMuon/python/selectors_cff.py
@@ -12,8 +12,24 @@ muonTPSet = cms.PSet(
     minRapidity = cms.double(-2.4),
     maxRapidity = cms.double(2.4),
     signalOnly = cms.bool(True),
-    intimeOnly = cms.bool(False),
-    stableOnly = cms.bool(False),
+    intimeOnly = cms.bool(True),  # discard OOT PU
+    stableOnly = cms.bool(True),  # discard decays in flight from the signal event
+    chargedOnly = cms.bool(True)
+)
+
+me0MuonTPSet = cms.PSet(
+    src = cms.InputTag("mix", "MergedTrackTruth"),
+    pdgId = cms.vint32(13, -13),
+    tip = cms.double(3.5),
+    lip = cms.double(30.0),
+    minHit = cms.int32(0),
+    ptMin = cms.double(0.9),
+    ptMax = cms.double(1e100),
+    minRapidity = cms.double(-2.8),
+    maxRapidity = cms.double(2.8),
+    signalOnly = cms.bool(True),
+    intimeOnly = cms.bool(True),  # discard OOT PU
+    stableOnly = cms.bool(True),  # discard decays in flight from the signal event
     chargedOnly = cms.bool(True)
 )
 
@@ -28,8 +44,23 @@ displacedMuonTPSet = cms.PSet(
     minRapidity = cms.double(-2.4),
     maxRapidity = cms.double(2.4),
     signalOnly = cms.bool(True),
-    intimeOnly = cms.bool(False),
-    stableOnly = cms.bool(False),
+    intimeOnly = cms.bool(True),  # discard OOT PU
+    stableOnly = cms.bool(True), # accept only TP from the Generator (linked to GenParticles)
+    chargedOnly = cms.bool(True)
+)
+
+cosmicMuonTPSet = cms.PSet(
+    src = cms.InputTag("mix", "MergedTrackTruth"),
+    pdgId = cms.vint32(13, -13),
+    tip = cms.double(85.),  # radius to have at least the 3 outermost TOB layers
+    lip = cms.double(210.), # z to have at least the 3 outermost TEC layers
+    minHit = cms.int32(0),
+    ptMin = cms.double(0.9),
+    ptMax = cms.double(1e100),
+    minRapidity = cms.double(-2.4),
+    maxRapidity = cms.double(2.4),
+    signalOnly = cms.bool(True),
+    stableOnly = cms.bool(True), # accept only TP from the Generator (linked to GenParticles)
     chargedOnly = cms.bool(True)
 )
