+++ Validation/RecoMuon/python/track_selectors_cff.py
@@ -2,36 +2,36 @@ import FWCore.ParameterSet.Config as cms
 
 # select probe tracks
 import PhysicsTools.RecoAlgos.recoTrackSelector_cfi
-NEWprobeTracks = PhysicsTools.RecoAlgos.recoTrackSelector_cfi.recoTrackSelector.clone()
-#NEWprobeTracks.quality = cms.vstring('highPurity')   # previous setting
-#NEWprobeTracks.quality = cms.vstring('loose')        # default
-NEWprobeTracks.tip = cms.double(3.5)
-NEWprobeTracks.lip = cms.double(30.)
-NEWprobeTracks.ptMin = cms.double(4.0)
-NEWprobeTracks.minRapidity = cms.double(-2.4)
-NEWprobeTracks.maxRapidity = cms.double(2.4)
-NEWprobeTracks_seq = cms.Sequence( NEWprobeTracks )
+probeTracks = PhysicsTools.RecoAlgos.recoTrackSelector_cfi.recoTrackSelector.clone()
+#probeTracks.quality = cms.vstring('highPurity')   # previous setting
+#probeTracks.quality = cms.vstring('loose')        # default
+probeTracks.tip = cms.double(3.5)
+probeTracks.lip = cms.double(30.)
+probeTracks.ptMin = cms.double(4.0)
+probeTracks.minRapidity = cms.double(-2.4)
+probeTracks.maxRapidity = cms.double(2.4)
+probeTracks_seq = cms.Sequence( probeTracks )
 
 import SimMuon.MCTruth.MuonTrackProducer_cfi
-NEWextractGemMuons = SimMuon.MCTruth.MuonTrackProducer_cfi.muonTrackProducer.clone()
-NEWextractGemMuons.selectionTags = ('All',)
-NEWextractGemMuons.trackType = "gemMuonTrack"
-NEWextractGemMuonsTracks_seq = cms.Sequence( NEWextractGemMuons )
+extractGemMuons = SimMuon.MCTruth.MuonTrackProducer_cfi.muonTrackProducer.clone()
+extractGemMuons.selectionTags = ('All',)
+extractGemMuons.trackType = "gemMuonTrack"
+extractGemMuonsTracks_seq = cms.Sequence( extractGemMuons )
 
-NEWextractMe0Muons = SimMuon.MCTruth.MuonTrackProducer_cfi.muonTrackProducer.clone()
-NEWextractMe0Muons.selectionTags = cms.vstring('All',)
-NEWextractMe0Muons.trackType = "me0MuonTrack"
-NEWextractMe0MuonsTracks_seq = cms.Sequence( NEWextractMe0Muons )
+extractMe0Muons = SimMuon.MCTruth.MuonTrackProducer_cfi.muonTrackProducer.clone()
+extractMe0Muons.selectionTags = cms.vstring('All',)
+extractMe0Muons.trackType = "me0MuonTrack"
+extractMe0MuonsTracks_seq = cms.Sequence( extractMe0Muons )
 
 #
 # Configuration for Seed track extractor
 #
 
 import SimMuon.MCTruth.SeedToTrackProducer_cfi
-NEWseedsOfSTAmuons = SimMuon.MCTruth.SeedToTrackProducer_cfi.SeedToTrackProducer.clone()
-NEWseedsOfSTAmuons.L2seedsCollection = cms.InputTag("ancientMuonSeed")
-NEWseedsOfSTAmuons_seq = cms.Sequence( NEWseedsOfSTAmuons )
+seedsOfSTAmuons = SimMuon.MCTruth.SeedToTrackProducer_cfi.SeedToTrackProducer.clone()
+seedsOfSTAmuons.L2seedsCollection = cms.InputTag("ancientMuonSeed")
+seedsOfSTAmuons_seq = cms.Sequence( seedsOfSTAmuons )
 
-NEWseedsOfDisplacedSTAmuons = SimMuon.MCTruth.SeedToTrackProducer_cfi.SeedToTrackProducer.clone()
-NEWseedsOfDisplacedSTAmuons.L2seedsCollection = cms.InputTag("displacedMuonSeeds")
-NEWseedsOfDisplacedSTAmuons_seq = cms.Sequence( NEWseedsOfDisplacedSTAmuons )
+seedsOfDisplacedSTAmuons = SimMuon.MCTruth.SeedToTrackProducer_cfi.SeedToTrackProducer.clone()
+seedsOfDisplacedSTAmuons.L2seedsCollection = cms.InputTag("displacedMuonSeeds")
+seedsOfDisplacedSTAmuons_seq = cms.Sequence( seedsOfDisplacedSTAmuons )
